refactor(ui): extract final 7 domain modules from app.js (task 149)
Extract adminUsers, dragDrop, shortcuts, commandPalette, taskDrawerAssist,
onCreateAssist (~1,250 lines), and todayPlan (~870 lines) into dedicated ES6
modules. Wire all cross-module calls through hooks object. Use namespace
imports for dragDrop, taskDrawerAssist, onCreateAssist, todayPlan to keep
the import section compact.

app.js reduced from 5,423 → 1,975 lines (63% reduction).
