+++ public/app.js
@@ -380,6 +380,31 @@ import {
   openDrawerDangerZone,
   bindTodoDrawerHandlers,
 } from "./drawerUi.js";
+import {
+  loadAdminUsers,
+  renderAdminUsers,
+  changeUserRole,
+  deleteUser,
+} from "./adminUsers.js";
+import * as DragDrop from "./dragDrop.js";
+import { toggleShortcuts, closeShortcutsOverlay } from "./shortcuts.js";
+import {
+  getCommandPaletteElements,
+  buildCommandPaletteItems,
+  getCommandPaletteCommandMatches,
+  getCommandPaletteTaskMatches,
+  getCommandPaletteRenderModel,
+  renderCommandPalette,
+  executeCommandPaletteItem,
+  moveCommandPaletteSelection,
+  closeCommandPalette,
+  openCommandPalette,
+  toggleCommandPalette,
+  bindCommandPaletteHandlers,
+} from "./commandPalette.js";
+import * as TaskDrawerAssist from "./taskDrawerAssist.js";
+import * as OnCreateAssist from "./onCreateAssist.js";
+import * as TodayPlan from "./todayPlan.js";
 
 // =============================================================================
 // EventBus — minimal pub-sub for decoupled state→render wiring.
@@ -398,70 +423,6 @@ const EventBus = (() => {
     },
   };
 })();
-
-// =============================================================================
-// TASK 140 DEPENDENCY GRAPH ANALYSIS
-// Generated before any split attempt. Status: BLOCKED (see below).
-//
-// PROPOSED MODULE BOUNDARIES:
-//   todosService.js  — CRUD operations for todos (loadTodos, addTodo, deleteTodo,
-//                      toggleTodo, updateTodo, deleteSelected, undo pipeline)
-//   projectsState.js — Project state (state.customProjects, state.projectRecords,
-//                      state.projectHeadingsByProjectId, setSelectedProjectKey,
-//                      loadProjects, createSubproject, renameProjectTree)
-//   drawerUi.js      — Todo drawer panel (openTodoDrawer, closeTodoDrawer,
-//                      renderTodoDrawer, syncTodoDrawerStateWithRender,
-//                      drawer draft save pipeline)
-//   filterLogic.js   — Filtering + rendering (filterTodos, applyFiltersAndRender,
-//                      renderTodos, updateCategoryFilter, setDateView, clearFilters)
-//   overlayManager.js — Modals + overlays (openEditTodoModal, closeEditTodoModal,
-//                       openProjectCrudModal, openProjectDeleteDialog,
-//                       toggleCommandPalette, toggleShortcuts)
-//
-// CONFIRMED CIRCULAR IMPORT CHAINS:
-//
-//   ① todosService.js ↔ filterLogic.js
-//      • loadTodos() / addTodo() / deleteTodo() / toggleTodo() → call filterTodos()
-//        (filterLogic.js must be imported by todosService.js)
-//      • filterTodos() / renderTodos() → read `todos` array
-//        (todosService.js must be imported by filterLogic.js for `todos` state)
-//      → CIRCULAR: todosService → filterLogic → todosService
-//
-//   ② projectsState.js ↔ filterLogic.js
-//      • setSelectedProjectKey() → calls applyFiltersAndRender()   [filterLogic.js]
-//      • applyFiltersAndRender() → renderTodos() reads state.customProjects,
-//        state.projectRecords, state.projectHeadingsByProjectId                [projectsState.js]
-//      → CIRCULAR: projectsState → filterLogic → projectsState
-//
-//   ③ drawerUi.js ↔ filterLogic.js
-//      • applyFiltersAndRender() → calls syncTodoDrawerStateWithRender() [drawerUi.js]
-//      • openTodoDrawer() / closeTodoDrawer() → call filterTodos()       [filterLogic.js]
-//      → CIRCULAR: filterLogic → drawerUi → filterLogic
-//
-// RESOLUTION ANALYSIS:
-//   Clean resolution requires ONE of:
-//   (A) A shared mutable state module (e.g. store.js) — NOT in the allowed file list.
-//   (B) Changing filterTodos() / setSelectedProjectKey() signatures to accept state
-//       as parameters — explicitly FORBIDDEN by task constraints.
-//   (C) Event-based decoupling (CustomEvent fire-and-forget) — changes the
-//       synchronous call semantics, constituting a behavior change (BLOCKED).
-//   (D) Dynamic import() to break cycles — requires async code paths where
-//       currently synchronous; constitutes a behavior change (BLOCKED).
-//   (E) Completely different module boundaries (e.g. co-locate todos state WITH
-//       filterLogic.js) — still leaves chain ③ unresolved.
-//
-// ADDITIONAL BLOCKER — window[functionName] dispatcher (line ~13318):
-//   The data-onclick dispatcher resolves ~90 handler functions via window[name].
-//   Switching to type="module" removes automatic global scope for all functions.
-//   Every handler would need an explicit window.xxx = xxx assignment in app.js,
-//   making app.js non-thin and defeating the entry-point-orchestrator goal.
-//
-// VERDICT: BLOCKED
-//   Per task constraint: "Circular import that cannot be resolved cleanly →
-//   BLOCKED, report to user."
-//   Three independent circular chains exist. None can be resolved without a
-//   shared store module (file not in allowed list) or forbidden signature changes.
-// =============================================================================
 
 // Configuration
 const API_URL =
@@ -699,11 +660,11 @@ const {
 const PROJECTS_RAIL_COLLAPSED_STORAGE_KEY = "todos:projects-rail-collapsed";
 const AI_WORKSPACE_COLLAPSED_STORAGE_KEY = "todos:ai-collapsed";
 const AI_WORKSPACE_VISIBLE_STORAGE_KEY = "todos:ai-visible";
-const AI_ON_CREATE_DISMISSED_STORAGE_KEY = "todos:ai-on-create-dismissed";
 const QUICK_ENTRY_PROPERTIES_OPEN_STORAGE_KEY =
   "todos:quick-entry-properties-open";
 const SIDEBAR_NAV_ITEMS = [];
-state.onCreateAssistState.dismissedTodoIds = loadOnCreateDismissedTodoIds();
+state.onCreateAssistState.dismissedTodoIds =
+  OnCreateAssist.loadOnCreateDismissedTodoIds();
 
 const HOME_TOP_FOCUS_CACHE_KEY = "todos:home-top-focus-cache";
 const HOME_TOP_FOCUS_CACHE_MAX_AGE_MS = 6 * 60 * 60 * 1000;
@@ -877,49 +838,6 @@ function reorderProjectHeadings(draggedId, targetId, placement = "before") {
   })();
 }
 
-function getHeadingDropTargetFromTodo(todoId, dropPosition = "before") {
-  const projectName = getSelectedProjectKey();
-  const headings = getProjectHeadings(projectName);
-  if (!headings.length) {
-    return null;
-  }
-  const headingIds = new Set(headings.map((heading) => String(heading.id)));
-  const todo = state.todos.find((item) => String(item.id) === String(todoId));
-  const todoHeadingId = String(todo?.headingId || "");
-  if (todoHeadingId && headingIds.has(todoHeadingId)) {
-    if (todoHeadingId === String(draggedHeadingId || "")) {
-      const currentIndex = headings.findIndex(
-        (heading) => String(heading.id) === todoHeadingId,
-      );
-      if (currentIndex >= 0) {
-        const previousHeading = headings[currentIndex - 1] || null;
-        const nextHeading = headings[currentIndex + 1] || null;
-        if (dropPosition === "before") {
-          if (previousHeading?.id) {
-            return { targetId: String(previousHeading.id), placement: "after" };
-          }
-          if (nextHeading?.id) {
-            return { targetId: String(nextHeading.id), placement: "before" };
-          }
-        } else {
-          if (nextHeading?.id) {
-            return { targetId: String(nextHeading.id), placement: "before" };
-          }
-          if (previousHeading?.id) {
-            return { targetId: String(previousHeading.id), placement: "after" };
-          }
-        }
-      }
-    }
-    return { targetId: todoHeadingId, placement: dropPosition };
-  }
-  const edgeHeading =
-    dropPosition === "after" ? headings[headings.length - 1] : headings[0];
-  if (!edgeHeading?.id) {
-    return null;
-  }
-  return { targetId: String(edgeHeading.id), placement: dropPosition };
-}
 async function moveTodoToHeading(todoId, headingIdValue) {
   const todo = state.todos.find((item) => item.id === todoId);
   if (!todo) return;
@@ -944,2876 +862,6 @@ async function moveTodoToHeading(todoId, headingIdValue) {
     console.error("Move todo heading failed:", error);
     showMessage("todosMessage", "Failed to move task to heading", "error");
   }
-}
-
-function getCommandPaletteElements() {
-  const overlay = document.getElementById("commandPaletteOverlay");
-  const panel = document.getElementById("commandPalettePanel");
-  const input = document.getElementById("commandPaletteInput");
-  const list = document.getElementById("commandPaletteList");
-  const empty = document.getElementById("commandPaletteEmpty");
-  const title = document.getElementById("commandPaletteTitle");
-  if (!(overlay instanceof HTMLElement)) return null;
-  if (!(panel instanceof HTMLElement)) return null;
-  if (!(input instanceof HTMLInputElement)) return null;
-  if (!(list instanceof HTMLElement)) return null;
-  if (!(empty instanceof HTMLElement)) return null;
-  if (!(title instanceof HTMLElement)) return null;
-  return { overlay, panel, input, list, empty, title };
-}
-
-function buildCommandPaletteItems() {
-  const baseItems = [
-    {
-      id: "add-task",
-      label: "Add task",
-      type: "action",
-      payload: "add-task",
-    },
-    {
-      id: "all-tasks",
-      label: "Go to All tasks",
-      type: "project",
-      payload: "",
-    },
-  ];
-
-  const projectItems = getAllProjects().map((projectName) => ({
-    id: `project-${projectName}`,
-    label: `Go to project: ${getProjectLeafName(projectName)}`,
-    type: "project",
-    payload: projectName,
-  }));
-
-  return [...baseItems, ...projectItems];
-}
-
-function getCommandPaletteCommandMatches(query) {
-  if (!query) {
-    return state.commandPaletteItems;
-  }
-  return state.commandPaletteItems.filter((item) =>
-    item.label.toLowerCase().includes(query),
-  );
-}
-
-function getCommandPaletteTaskMatches(query) {
-  if (!query) {
-    return [];
-  }
-
-  const normalizedQuery = query.toLowerCase();
-  const ranked = state.todos
-    .map((todo) => {
-      const title = String(todo.title || "");
-      const description = String(todo.description || "");
-      const titleLower = title.toLowerCase();
-      const descriptionLower = description.toLowerCase();
-
-      let score = -1;
-      if (titleLower.startsWith(normalizedQuery)) {
-        score = 0;
-      } else if (titleLower.includes(normalizedQuery)) {
-        score = 1;
-      } else if (descriptionLower.includes(normalizedQuery)) {
-        score = 2;
-      }
-
-      if (score === -1) return null;
-
-      const dueAt = todo.dueDate ? new Date(todo.dueDate).getTime() : Infinity;
-      return {
-        id: `task-${todo.id}`,
-        type: "task",
-        todoId: String(todo.id),
-        label: title,
-        score,
-        dueAt: Number.isFinite(dueAt) ? dueAt : Infinity,
-        completed: !!todo.completed,
-        meta: [
-          todo.category ? `Project: ${todo.category}` : "",
-          todo.dueDate
-            ? `Due: ${new Date(todo.dueDate).toLocaleDateString()}`
-            : "",
-        ]
-          .filter(Boolean)
-          .join(" • "),
-      };
-    })
-    .filter(Boolean)
-    .sort((a, b) => {
-      if (a.score !== b.score) return a.score - b.score;
-      if (a.dueAt !== b.dueAt) return a.dueAt - b.dueAt;
-      const titleCompare = a.label.localeCompare(b.label);
-      if (titleCompare !== 0) return titleCompare;
-      return a.todoId.localeCompare(b.todoId);
-    });
-
-  return ranked.slice(0, 6);
-}
-
-function getCommandPaletteRenderModel() {
-  const query = state.commandPaletteQuery.trim().toLowerCase();
-  const commandMatches = getCommandPaletteCommandMatches(query);
-  const taskMatches = getCommandPaletteTaskMatches(query);
-  const rows = [];
-
-  if (commandMatches.length > 0) {
-    rows.push({
-      kind: "section",
-      id: "commands-section",
-      label: "Commands",
-    });
-    commandMatches.forEach((item) => {
-      rows.push({ kind: "item", item });
-    });
-  }
-
-  if (query) {
-    rows.push({
-      kind: "section",
-      id: "tasks-section",
-      label: "Tasks",
-    });
-    if (taskMatches.length === 0) {
-      rows.push({
-        kind: "empty",
-        id: "tasks-empty",
-        label: "No tasks found",
-      });
-    } else {
-      taskMatches.forEach((item) => {
-        rows.push({ kind: "item", item });
-      });
-    }
-  }
-
-  const selectableItems = rows
-    .filter((row) => row.kind === "item")
-    .map((row) => row.item);
-
-  const hasAnyResults =
-    commandMatches.length > 0 || (!!query && taskMatches.length > 0);
-  return { rows, selectableItems, hasAnyResults };
-}
-
-function renderCommandPalette() {
-  const refs = getCommandPaletteElements();
-  if (!refs) return;
-
-  refs.overlay.classList.toggle(
-    "command-palette-overlay--open",
-    state.isCommandPaletteOpen,
-  );
-  refs.overlay.setAttribute("aria-hidden", String(!state.isCommandPaletteOpen));
-  refs.input.value = state.commandPaletteQuery;
-
-  const renderModel = getCommandPaletteRenderModel();
-  state.commandPaletteSelectableItems = renderModel.selectableItems;
-  if (state.commandPaletteSelectableItems.length === 0) {
-    state.commandPaletteIndex = 0;
-  } else if (
-    state.commandPaletteIndex >
-    state.commandPaletteSelectableItems.length - 1
-  ) {
-    state.commandPaletteIndex = state.commandPaletteSelectableItems.length - 1;
-  }
-
-  refs.input.setAttribute("aria-expanded", String(state.isCommandPaletteOpen));
-  refs.input.setAttribute(
-    "aria-activedescendant",
-    state.commandPaletteSelectableItems.length > 0
-      ? `commandPaletteOption-${state.commandPaletteIndex}`
-      : "",
-  );
-
-  let selectableIndex = -1;
-  refs.list.innerHTML = renderModel.rows
-    .map((row) => {
-      if (row.kind === "section") {
-        return `<div class="command-palette-section" role="presentation">${escapeHtml(row.label)}</div>`;
-      }
-      if (row.kind === "empty") {
-        return `<div class="command-palette-inline-empty" role="status">${escapeHtml(row.label)}</div>`;
-      }
-
-      selectableIndex += 1;
-      const isActive = selectableIndex === state.commandPaletteIndex;
-      const item = row.item;
-      if (item.type === "task") {
-        return `
-          <button
-            type="button"
-            id="commandPaletteOption-${selectableIndex}"
-            class="command-palette-option command-palette-option--task ${isActive ? "command-palette-option--active" : ""} ${item.completed ? "command-palette-option--completed" : ""}"
-            role="option"
-            aria-selected="${isActive ? "true" : "false"}"
-            data-command-index="${selectableIndex}"
-            data-command-id="${escapeHtml(item.id)}"
-          >
-            <span class="command-palette-option__title">${escapeHtml(item.label)}</span>
-            <span class="command-palette-option__meta">${escapeHtml(item.meta || (item.completed ? "Completed" : ""))}</span>
-          </button>
-        `;
-      }
-
-      return `
-        <button
-          type="button"
-          id="commandPaletteOption-${selectableIndex}"
-          class="command-palette-option ${isActive ? "command-palette-option--active" : ""}"
-          role="option"
-          aria-selected="${isActive ? "true" : "false"}"
-          data-command-index="${selectableIndex}"
-          data-command-id="${escapeHtml(item.id)}"
-        >
-          ${escapeHtml(item.label)}
-        </button>
-      `;
-    })
-    .join("");
-
-  refs.empty.hidden = renderModel.hasAnyResults;
-}
-
-function executeCommandPaletteItem(item, triggerEl = null) {
-  if (!item) return;
-
-  const todosView = document.getElementById("todosView");
-  const shouldSwitchToTodos =
-    !(todosView instanceof HTMLElement) ||
-    !todosView.classList.contains("active") ||
-    todosView.classList.contains("todos-view--settings-active");
-
-  if (item.type === "action" && item.payload === "add-task") {
-    const todosTab = document.querySelector(
-      ".nav-tab[data-onclick*=\"switchView('todos'\"]",
-    );
-    if (shouldSwitchToTodos) {
-      switchView("todos", todosTab instanceof HTMLElement ? todosTab : null);
-    }
-    closeCommandPalette({ restoreFocus: false });
-    window.requestAnimationFrame(() => {
-      openTaskComposer(triggerEl);
-    });
-    return;
-  }
-
-  if (item.type === "project") {
-    const todosTab = document.querySelector(
-      ".nav-tab[data-onclick*=\"switchView('todos'\"]",
-    );
-    if (shouldSwitchToTodos) {
-      switchView("todos", todosTab instanceof HTMLElement ? todosTab : null);
-    }
-    state.currentWorkspaceView = item.payload ? "project" : "all";
-    clearHomeListDrilldown();
-    setSelectedProjectKey(String(item.payload || ""));
-    closeCommandPalette({ restoreFocus: false });
-  }
-
-  if (item.type === "task") {
-    const todosTab = document.querySelector(
-      ".nav-tab[data-onclick*=\"switchView('todos'\"]",
-    );
-    if (shouldSwitchToTodos) {
-      switchView("todos", todosTab instanceof HTMLElement ? todosTab : null);
-    }
-
-    closeCommandPalette({ restoreFocus: false });
-    window.requestAnimationFrame(() => {
-      openTodoDrawer(
-        item.todoId,
-        triggerEl instanceof HTMLElement ? triggerEl : null,
-      );
-    });
-  }
-}
-
-function moveCommandPaletteSelection(delta) {
-  const visibleCount = state.commandPaletteSelectableItems.length;
-  if (visibleCount === 0) {
-    state.commandPaletteIndex = 0;
-    renderCommandPalette();
-    return;
-  }
-  state.commandPaletteIndex =
-    (state.commandPaletteIndex + delta + visibleCount) % visibleCount;
-  renderCommandPalette();
-}
-
-function closeCommandPalette({ restoreFocus = true } = {}) {
-  if (!state.isCommandPaletteOpen) return;
-  state.isCommandPaletteOpen = false;
-  state.commandPaletteQuery = "";
-  state.commandPaletteIndex = 0;
-  state.commandPaletteSelectableItems = [];
-  renderCommandPalette();
-  DialogManager.close("commandPalette");
-
-  if (restoreFocus && state.lastFocusedBeforePalette instanceof HTMLElement) {
-    state.lastFocusedBeforePalette.focus({ preventScroll: true });
-  }
-}
-
-function openCommandPalette() {
-  if (!state.currentUser) return;
-
-  const refs = getCommandPaletteElements();
-  if (!refs) return;
-  if (state.isCommandPaletteOpen) return;
-
-  state.lastFocusedBeforePalette =
-    document.activeElement instanceof HTMLElement
-      ? document.activeElement
-      : null;
-  state.commandPaletteItems = buildCommandPaletteItems();
-  state.commandPaletteSelectableItems = [];
-  state.commandPaletteQuery = "";
-  state.commandPaletteIndex = 0;
-  state.isCommandPaletteOpen = true;
-  renderCommandPalette();
-  DialogManager.open("commandPalette", refs.overlay, {
-    onEscape: () => closeCommandPalette({ restoreFocus: true }),
-    backdrop: false,
-  });
-
-  window.requestAnimationFrame(() => {
-    refs.input.focus();
-    refs.input.select();
-  });
-}
-
-function toggleCommandPalette() {
-  if (state.isCommandPaletteOpen) {
-    closeCommandPalette({ restoreFocus: true });
-    return;
-  }
-  openCommandPalette();
-}
-
-function taskDrawerDismissKey(todoId) {
-  return `taskDrawerAssist:dismissed:${todoId}`;
-}
-
-function getTaskDrawerSuggestionLabel(type) {
-  return labelForType(type);
-}
-
-function normalizeTaskDrawerAssistEnvelope(
-  aiSuggestionId,
-  outputEnvelope,
-  fallbackTodoId,
-) {
-  const suggestionsRaw = Array.isArray(outputEnvelope?.suggestions)
-    ? outputEnvelope.suggestions
-    : [];
-  const normalized = suggestionsRaw
-    .map((item, index) => {
-      if (!item || typeof item !== "object") return null;
-      const suggestion = item;
-      const type = String(suggestion.type || "");
-      if (!shouldRenderTypeForSurface("task_drawer", type)) {
-        return null;
-      }
-      const payload =
-        suggestion.payload && typeof suggestion.payload === "object"
-          ? { ...suggestion.payload }
-          : {};
-      if (typeof payload.todoId !== "string" || !payload.todoId.trim()) {
-        payload.todoId = fallbackTodoId;
-      }
-      return {
-        type,
-        suggestionId:
-          String(suggestion.suggestionId || "").trim() ||
-          `${aiSuggestionId || "drawer"}-${index + 1}`,
-        confidence: Math.max(
-          0,
-          Math.min(1, Number(suggestion.confidence) || 0),
-        ),
-        rationale: truncateRationale(suggestion.rationale, 120),
-        payload,
-        requiresConfirmation: needsConfirmation(suggestion),
-        dismissed: false,
-      };
-    })
-    .filter(Boolean);
-  return capSuggestions(sortSuggestions("task_drawer", normalized), 6);
-}
-
-function renderTaskDrawerSuggestionSummary(suggestion) {
-  if (suggestion.type === "rewrite_title") {
-    return String(suggestion.payload.title || "Rewrite task title");
-  }
-  if (suggestion.type === "set_due_date") {
-    const due = new Date(String(suggestion.payload.dueDateISO || ""));
-    return Number.isNaN(due.getTime())
-      ? "Set due date"
-      : `Set due ${due.toLocaleDateString()}`;
-  }
-  if (suggestion.type === "set_priority") {
-    return `Set priority ${String(suggestion.payload.priority || "").toUpperCase()}`;
-  }
-  if (suggestion.type === "set_project") {
-    return `Set project ${String(suggestion.payload.projectName || suggestion.payload.category || "")}`.trim();
-  }
-  if (suggestion.type === "set_category") {
-    return `Set category ${String(suggestion.payload.category || "")}`.trim();
-  }
-  if (suggestion.type === "split_subtasks") {
-    const count = Array.isArray(suggestion.payload.subtasks)
-      ? suggestion.payload.subtasks.length
-      : 0;
-    return `Add ${count} subtasks`;
-  }
-  if (suggestion.type === "propose_next_action") {
-    return String(suggestion.payload.text || suggestion.payload.title || "");
-  }
-  if (suggestion.type === "ask_clarification") {
-    return String(suggestion.payload.question || "Need one clarification");
-  }
-  return "Suggestion";
-}
-
-function renderTaskDrawerAssistSection(todoId) {
-  // Lint-first gate: show lint chip only when a heuristic fires.
-  // Full panel should render only after explicit user action.
-  // In debug mode, always show the full panel so developers see all metadata.
-  const assistState = state.taskDrawerAssistState;
-  if (!assistState.showFullAssist && !AI_DEBUG_ENABLED) {
-    const todo = getTodoById(todoId);
-    const issue = todo
-      ? lintTodoFields({
-          title: todo.title,
-          dueDate: todo.dueDate || "",
-          priority: todo.priority || "",
-          subtasks: Array.isArray(todo.subtasks) ? todo.subtasks : [],
-          allTodos: state.todos,
-        })
-      : null;
-    if (issue) {
-      // A lint issue is present — show only the chip (suppress full panel).
-      return `
-        <div class="todo-drawer__section">
-          <div class="todo-drawer__section-title">AI Suggestions</div>
-          ${renderLintChip(issue)}
-        </div>
-      `;
-    }
-    return "";
-  }
-
-  if (!FEATURE_TASK_DRAWER_DECISION_ASSIST) {
-    return `
-      <div class="todo-drawer__section">
-        <div class="todo-drawer__section-title">AI Suggestions</div>
-        <div class="ai-empty" role="status">AI Suggestions unavailable.</div>
-      </div>
-    `;
-  }
-  const base = `
-    <div class="todo-drawer__section">
-      <div class="todo-drawer__section-title">AI Suggestions</div>
-      ${renderAiDebugMeta({
-        requestId: assistState.requestId,
-        generatedAt: assistState.generatedAt,
-        contractVersion: assistState.contractVersion,
-      })}
-      ${assistState.loading ? '<div class="ai-empty" role="status">Loading suggestions...</div>' : ""}
-      ${
-        assistState.unavailable
-          ? '<div class="ai-empty" role="status">AI Suggestions unavailable.</div>'
-          : ""
-      }
-      ${assistState.error ? `<div class="ai-empty" role="status">${escapeHtml(assistState.error)}</div>` : ""}
-      ${
-        !assistState.loading &&
-        !assistState.unavailable &&
-        !assistState.error &&
-        (assistState.mustAbstain || assistState.suggestions.length === 0)
-          ? '<div class="ai-empty" role="status">No suggestions right now.</div>'
-          : ""
-      }
-      ${
-        assistState.suggestions.length > 0
-          ? `<div class="todo-drawer-ai-list">
-              ${assistState.suggestions
-                .map((suggestion) => {
-                  const confidenceLabel = confidenceLabelForSuggestion(
-                    suggestion.confidence,
-                  );
-                  const applying =
-                    assistState.applyingSuggestionId ===
-                    suggestion.suggestionId;
-                  const confirmOpen =
-                    assistState.confirmSuggestionId === suggestion.suggestionId;
-                  return `
-                    <div class="todo-drawer-ai-card ai-card" data-testid="task-drawer-ai-card-${escapeHtml(suggestion.suggestionId)}">
-                      <div class="todo-drawer-ai-card__top">
-                        <span class="todo-drawer-ai-card__label">${escapeHtml(getTaskDrawerSuggestionLabel(suggestion.type))}</span>
-                        <span class="ai-badge ai-badge--${escapeHtml(confidenceBand(suggestion.confidence))}" aria-label="Confidence ${escapeHtml(confidenceLabel)}">${escapeHtml(confidenceLabel)}</span>
-                      </div>
-                      <div class="todo-drawer-ai-card__summary">${escapeHtml(renderTaskDrawerSuggestionSummary(suggestion))}</div>
-                      <div class="todo-drawer-ai-card__rationale ai-tooltip">${escapeHtml(suggestion.rationale)}</div>
-                      ${renderAiDebugSuggestionId(suggestion.suggestionId)}
-                      ${
-                        assistState.lastUndoSuggestionId ===
-                        suggestion.suggestionId
-                          ? `<div class="ai-tooltip">Undone locally</div>`
-                          : ""
-                      }
-                      <div class="todo-drawer-ai-card__actions ai-actions">
-                        ${
-                          confirmOpen
-                            ? `
-                              <div class="ai-confirm">
-                                <button
-                                  type="button"
-                                  class="ai-action-btn"
-                                  data-testid="task-drawer-ai-confirm-${escapeHtml(suggestion.suggestionId)}"
-                                  data-drawer-ai-action="confirm-apply"
-                                  data-drawer-ai-suggestion-id="${escapeHtml(suggestion.suggestionId)}"
-                                  aria-label="Confirm apply ${escapeHtml(getTaskDrawerSuggestionLabel(suggestion.type))}"
-                                >
-                                  Confirm
-                                </button>
-                                <button
-                                  type="button"
-                                  class="ai-action-btn"
-                                  data-testid="task-drawer-ai-cancel-${escapeHtml(suggestion.suggestionId)}"
-                                  data-drawer-ai-action="cancel-confirm"
-                                  data-drawer-ai-suggestion-id="${escapeHtml(suggestion.suggestionId)}"
-                                  aria-label="Cancel confirmation"
-                                >
-                                  Cancel
-                                </button>
-                              </div>
-                            `
-                            : `
-                              <button
-                                type="button"
-                                class="ai-action-btn"
-                                data-testid="task-drawer-ai-apply-${escapeHtml(suggestion.suggestionId)}"
-                                data-drawer-ai-action="apply"
-                                data-drawer-ai-suggestion-id="${escapeHtml(suggestion.suggestionId)}"
-                                aria-label="Apply ${escapeHtml(getTaskDrawerSuggestionLabel(suggestion.type))}"
-                                ${applying ? "disabled" : ""}
-                              >
-                                ${applying ? "Applying..." : "Apply"}
-                              </button>
-                            `
-                        }
-                        <button
-                          type="button"
-                          class="ai-action-btn"
-                          data-testid="task-drawer-ai-dismiss-${escapeHtml(suggestion.suggestionId)}"
-                          data-drawer-ai-action="dismiss"
-                          data-drawer-ai-suggestion-id="${escapeHtml(suggestion.suggestionId)}"
-                          aria-label="Dismiss ${escapeHtml(getTaskDrawerSuggestionLabel(suggestion.type))}"
-                        >
-                          Dismiss
-                        </button>
-                        ${
-                          assistState.undoBySuggestionId[
-                            suggestion.suggestionId
-                          ]
-                            ? `
-                              <button
-                                type="button"
-                                class="ai-undo"
-                                data-testid="task-drawer-ai-undo-${escapeHtml(suggestion.suggestionId)}"
-                                data-drawer-ai-action="undo"
-                                data-drawer-ai-suggestion-id="${escapeHtml(suggestion.suggestionId)}"
-                                aria-label="Undo ${escapeHtml(getTaskDrawerSuggestionLabel(suggestion.type))}"
-                              >
-                                Undo
-                              </button>
-                            `
-                            : ""
-                        }
-                      </div>
-                    </div>
-                  `;
-                })
-                .join("")}
-            </div>`
-          : ""
-      }
-    </div>
-  `;
-
-  window.requestAnimationFrame(() => {
-    const confirmButton = document.querySelector(
-      `[data-drawer-ai-action="confirm-apply"][data-drawer-ai-suggestion-id="${escapeSelectorValue(
-        assistState.confirmSuggestionId || "",
-      )}"]`,
-    );
-    if (confirmButton instanceof HTMLElement) {
-      confirmButton.focus({ preventScroll: true });
-    }
-  });
-
-  return base;
-}
-
-function confidenceLabelForSuggestion(confidence) {
-  return confidenceLabel(confidence);
-}
-
-function getTodoById(todoId) {
-  return state.todos.find((todo) => todo.id === todoId) || null;
-}
-
-function getCurrentDrawerDraft(todo) {
-  if (!todo) return null;
-  if (!state.drawerDraft || state.drawerDraft.id !== todo.id) {
-    initializeDrawerDraft(todo);
-  }
-  return state.drawerDraft;
-}
-
-function captureDrawerFocusState() {
-  const refs = getTodoDrawerElements();
-  if (!refs) return null;
-  const active = document.activeElement;
-  if (!(active instanceof HTMLElement)) return null;
-  if (!refs.drawer.contains(active)) return null;
-  if (!active.id) return null;
-
-  const state = { id: active.id, selectionStart: null, selectionEnd: null };
-  if (
-    active instanceof HTMLInputElement ||
-    active instanceof HTMLTextAreaElement
-  ) {
-    state.selectionStart = active.selectionStart;
-    state.selectionEnd = active.selectionEnd;
-  }
-  return state;
-}
-
-function restoreDrawerFocusState(focusState) {
-  if (!focusState || !focusState.id) return;
-  const target = document.getElementById(focusState.id);
-  if (!(target instanceof HTMLElement)) return;
-
-  target.focus({ preventScroll: true });
-  if (
-    target instanceof HTMLInputElement ||
-    target instanceof HTMLTextAreaElement
-  ) {
-    if (
-      typeof focusState.selectionStart === "number" &&
-      typeof focusState.selectionEnd === "number"
-    ) {
-      target.setSelectionRange(
-        focusState.selectionStart,
-        focusState.selectionEnd,
-      );
-    }
-  }
-}
-
-function isMobileDrawerViewport() {
-  if (window.matchMedia) {
-    return window.matchMedia(MOBILE_DRAWER_MEDIA_QUERY).matches;
-  }
-  return window.innerWidth <= 768;
-}
-
-function updateDrawerDraftField(field, value) {
-  if (!state.drawerDraft) return;
-  state.drawerDraft[field] = value;
-  if (state.drawerSaveState !== "saving") {
-    setDrawerSaveState("idle");
-  }
-}
-
-function renderDrawerSubtasks(todo) {
-  if (!Array.isArray(todo.subtasks) || todo.subtasks.length === 0) {
-    return '<p class="todo-drawer__subtasks-empty">No subtasks</p>';
-  }
-
-  return `
-    <ul class="todo-drawer__subtasks-list">
-      ${todo.subtasks
-        .map((subtask) => {
-          const title = escapeHtml(String(subtask?.title || ""));
-          return `
-            <li class="todo-drawer__subtasks-item ${subtask?.completed ? "completed" : ""}">
-              <span aria-hidden="true">${subtask?.completed ? "✓" : "○"}</span>
-              <span>${title}</span>
-            </li>
-          `;
-        })
-        .join("")}
-    </ul>
-  `;
-  updateAiWorkspaceStatusChip();
-}
-
-function buildDrawerProjectOptions(selectedProject = "") {
-  const projects = getAllProjects();
-  return `<option value="">None</option>${projects
-    .map((project) => renderProjectOptionEntry(project, selectedProject))
-    .join("")}`;
-}
-
-async function fetchTaskDrawerLatestSuggestion(todoId) {
-  const response = await apiCall(
-    `${API_URL}/ai/suggestions/latest?todoId=${encodeURIComponent(todoId)}&surface=task_drawer`,
-  );
-  return response;
-}
-
-async function generateTaskDrawerSuggestion(todo) {
-  return apiCall(`${API_URL}/ai/decision-assist/stub`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({
-      surface: "task_drawer",
-      todoId: todo.id,
-      title: todo.title,
-      description: todo.description || "",
-      notes: todo.notes || "",
-    }),
-  });
-}
-
-function escapeSelectorValue(value) {
-  const raw = String(value);
-  if (window.CSS && typeof window.CSS.escape === "function") {
-    return window.CSS.escape(raw);
-  }
-  return raw.replace(/["\\]/g, "\\$&");
-}
-
-function renderTodoChips(todo, { isOverdue, dueDateStr }) {
-  const chips = [];
-
-  if (todo.dueDate) {
-    const dueLabel = `${isOverdue ? "⚠️" : "📅"} ${dueDateStr}`;
-    chips.push(
-      `<span class="todo-chip todo-chip--due ${isOverdue ? "todo-chip--due-overdue" : ""}" title="${escapeHtml(dueLabel)}">${escapeHtml(dueLabel)}</span>`,
-    );
-  }
-
-  if (todo.category) {
-    chips.push(
-      `<span class="todo-chip todo-chip--project" title="${escapeHtml(todo.category)}">🏷️ ${escapeHtml(todo.category)}</span>`,
-    );
-  }
-
-  if (todo.priority === "high" && chips.length < 2) {
-    chips.push(
-      `<span class="todo-chip todo-chip--priority priority-badge high" title="High priority">HIGH</span>`,
-    );
-  }
-
-  return chips.slice(0, 2).join("");
-}
-
-function resetOnCreateAssistState() {
-  const dismissedTodoIds =
-    state.onCreateAssistState?.dismissedTodoIds || new Set();
-  state.onCreateAssistState = createInitialOnCreateAssistState();
-  state.onCreateAssistState.dismissedTodoIds = dismissedTodoIds;
-}
-
-function loadOnCreateDismissedTodoIds() {
-  try {
-    const raw = window.localStorage.getItem(AI_ON_CREATE_DISMISSED_STORAGE_KEY);
-    const parsed = raw ? JSON.parse(raw) : [];
-    if (!Array.isArray(parsed)) return new Set();
-    return new Set(
-      parsed.map((value) => String(value || "").trim()).filter(Boolean),
-    );
-  } catch {
-    return new Set();
-  }
-}
-
-function persistOnCreateDismissedTodoIds() {
-  try {
-    const values = Array.from(
-      state.onCreateAssistState.dismissedTodoIds || new Set(),
-    );
-    window.localStorage.setItem(
-      AI_ON_CREATE_DISMISSED_STORAGE_KEY,
-      JSON.stringify(values),
-    );
-  } catch {
-    // Ignore storage failures.
-  }
-}
-
-function markOnCreateDismissed(todoId) {
-  if (!todoId) return;
-  state.onCreateAssistState.dismissedTodoIds.add(String(todoId));
-  persistOnCreateDismissedTodoIds();
-}
-
-function clearOnCreateDismissed(todoId) {
-  if (!todoId) return;
-  state.onCreateAssistState.dismissedTodoIds.delete(String(todoId));
-  persistOnCreateDismissedTodoIds();
-}
-
-function isOnCreateDismissed(todoId) {
-  if (!todoId) return false;
-  return state.onCreateAssistState.dismissedTodoIds.has(String(todoId));
-}
-
-function getOnCreateImpactRank(type) {
-  return impactRankForSurface(ON_CREATE_SURFACE, type);
-}
-
-function getOnCreateConfidenceBadge(confidence) {
-  return confidenceLabel(confidence);
-}
-
-function clampOnCreateRationale(value) {
-  return truncateRationale(value, 120);
-}
-
-function formatOnCreateSuggestionLabel(type) {
-  return labelForType(type);
-}
-
-function formatOnCreateChoiceValue(choice) {
-  if (typeof choice === "string") {
-    return { value: choice, label: choice };
-  }
-  const value = String(
-    choice?.value ||
-      choice?.projectName ||
-      choice?.category ||
-      choice?.label ||
-      "",
-  ).trim();
-  const label = String(choice?.label || value).trim();
-  if (!value) return null;
-  return {
-    value,
-    label,
-    projectName: choice?.projectName,
-    category: choice?.category,
-  };
-}
-
-function normalizeOnCreateSuggestion(rawSuggestion) {
-  if (!rawSuggestion || typeof rawSuggestion !== "object") return null;
-  const type = String(rawSuggestion.type || "");
-  if (!isKnownSuggestionType(type)) return null;
-  if (!shouldRenderTypeForSurface(ON_CREATE_SURFACE, type)) return null;
-  const suggestionId = String(rawSuggestion.suggestionId || "").trim();
-  if (!suggestionId) return null;
-  const payload =
-    rawSuggestion.payload && typeof rawSuggestion.payload === "object"
-      ? rawSuggestion.payload
-      : {};
-  const normalized = {
-    type,
-    suggestionId,
-    confidence: Math.max(0, Math.min(1, Number(rawSuggestion.confidence) || 0)),
-    rationale: clampOnCreateRationale(rawSuggestion.rationale),
-    payload,
-    requiresConfirmation: !!rawSuggestion.requiresConfirmation,
-    dismissed: false,
-    applied: false,
-    confirmationOpen: false,
-    undoSnapshot: null,
-    clarificationExpanded: false,
-    clarificationAnswered: false,
-    clarificationAnswer: "",
-    helperText: "",
-  };
-  if (type === "ask_clarification") {
-    const choicesRaw = Array.isArray(payload.choices) ? payload.choices : [];
-    normalized.payload = {
-      ...payload,
-      question: String(payload.question || "Pick one option").slice(0, 120),
-      choices: choicesRaw
-        .map((choice) => formatOnCreateChoiceValue(choice))
-        .filter(Boolean)
-        .slice(0, 4),
-    };
-  }
-  return normalized;
-}
-
-function buildOnCreateSuggestion(
-  type,
-  suggestionId,
-  confidence,
-  rationale,
-  payload,
-  options = {},
-) {
-  return {
-    type,
-    suggestionId,
-    confidence,
-    rationale,
-    payload,
-    ...(options.requiresConfirmation ? { requiresConfirmation: true } : {}),
-  };
-}
-
-function nextWeekdayAtNoonIso(day) {
-  const target = new Date();
-  const currentDay = target.getDay();
-  let delta = (day - currentDay + 7) % 7;
-  if (delta === 0) delta = 7;
-  target.setDate(target.getDate() + delta);
-  target.setHours(12, 0, 0, 0);
-  return target.toISOString();
-}
-
-function yesterdayAtNoonIso() {
-  const target = new Date();
-  target.setDate(target.getDate() - 1);
-  target.setHours(12, 0, 0, 0);
-  return target.toISOString();
-}
-
-function buildMockOnCreateAssistEnvelope(rawTitle) {
-  const title = String(rawTitle || "").trim();
-  const titleLower = title.toLowerCase();
-  const suggestions = [];
-
-  if (titleLower.includes("tomorrow")) {
-    const due = new Date();
-    due.setDate(due.getDate() + 1);
-    due.setHours(12, 0, 0, 0);
-    suggestions.push(
-      buildOnCreateSuggestion(
-        "set_due_date",
-        "oc-set-due-tomorrow",
-        0.88,
-        "Detected a relative deadline in the title.",
-        { dueDateISO: due.toISOString() },
-      ),
-    );
-  } else if (titleLower.includes("by friday")) {
-    suggestions.push(
-      buildOnCreateSuggestion(
-        "set_due_date",
-        "oc-set-due-friday",
-        0.82,
-        '"By Friday" implies a firm due date.',
-        { dueDateISO: nextWeekdayAtNoonIso(5) },
-      ),
-    );
-  }
-
-  if (titleLower.includes("yesterday")) {
-    suggestions.push(
-      buildOnCreateSuggestion(
-        "set_due_date",
-        "oc-set-due-past",
-        0.51,
-        "Parsed a past date mention. Confirm before applying.",
-        { dueDateISO: yesterdayAtNoonIso() },
-        { requiresConfirmation: true },
-      ),
-    );
-  }
-
-  if (/\burgent\b/.test(titleLower)) {
-    suggestions.push(
-      buildOnCreateSuggestion(
-        "set_priority",
-        "oc-set-priority-urgent",
-        0.9,
-        "“Urgent” strongly signals high priority.",
-        { priority: "high" },
-      ),
-    );
-  } else if (/\basap\b/.test(titleLower)) {
-    suggestions.push(
-      buildOnCreateSuggestion(
-        "set_priority",
-        "oc-set-priority-asap",
-        0.67,
-        "ASAP often maps to high priority. Confirm escalation.",
-        { priority: "high" },
-        { requiresConfirmation: true },
-      ),
-    );
-  }
-
-  const mentionsWebsite = titleLower.includes("website");
-  const mentionsMarketing = titleLower.includes("marketing");
-  if (mentionsWebsite && mentionsMarketing) {
-    suggestions.push(
-      buildOnCreateSuggestion(
-        "ask_clarification",
-        "oc-clarify-project",
-        0.6,
-        "Project target is ambiguous between website and marketing.",
-        {
-          questionId: "oc-project-choice",
-          question: "Which project should this task belong to?",
-          choices: [
-            { value: "Website", label: "Website", projectName: "Website" },
-            { value: "Marketing", label: "Marketing", category: "Marketing" },
-          ],
-        },
-      ),
-    );
-  } else if (mentionsWebsite) {
-    suggestions.push(
-      buildOnCreateSuggestion(
-        "set_project",
-        "oc-set-project-website",
-        0.84,
-        "Detected a known project keyword.",
-        { projectName: "Website", category: "Website" },
-      ),
-    );
-  } else if (mentionsMarketing) {
-    suggestions.push(
-      buildOnCreateSuggestion(
-        "set_category",
-        "oc-set-category-marketing",
-        0.8,
-        "Detected a category keyword.",
-        { category: "Marketing" },
-      ),
-    );
-  }
-
-  if (titleLower.includes("personal")) {
-    suggestions.push(
-      buildOnCreateSuggestion(
-        "set_category",
-        "oc-set-category-personal",
-        0.73,
-        "“Personal” maps cleanly to a category.",
-        { category: "Personal" },
-      ),
-    );
-  }
-
-  if (/\b(email|follow up|stuff)\b/.test(titleLower)) {
-    suggestions.push(
-      buildOnCreateSuggestion(
-        "rewrite_title",
-        "oc-rewrite-vague",
-        0.77,
-        "Title is vague; propose a concrete next-action title.",
-        { title: "Email stakeholder with specific next step and deadline" },
-      ),
-    );
-  }
-
-  if (titleLower.includes("unknown")) {
-    suggestions.push({
-      type: "not_supported",
-      suggestionId: "oc-unknown-type",
-      confidence: 0.5,
-      rationale: "Should be ignored by UI.",
-      payload: {},
-    });
-  }
-
-  return {
-    contractVersion: 1,
-    generatedAt: new Date().toISOString(),
-    surface: ON_CREATE_SURFACE,
-    requestId: `on-create-${encodeURIComponent(titleLower.slice(0, 24) || "empty")}`,
-    must_abstain: false,
-    suggestions,
-  };
-}
-
-function getOnCreateAssistElements() {
-  const row = document.getElementById("aiOnCreateAssistRow");
-  const titleInput = document.getElementById("todoInput");
-  const projectSelect = document.getElementById("todoProjectSelect");
-  const dueDateInput = document.getElementById("todoDueDateInput");
-  if (!(row instanceof HTMLElement)) return null;
-  if (!(titleInput instanceof HTMLInputElement)) return null;
-  return { row, titleInput, projectSelect, dueDateInput };
-}
-
-function ensureOnCreateProjectOption(projectName) {
-  const refs = getOnCreateAssistElements();
-  if (!refs || !(refs.projectSelect instanceof HTMLSelectElement)) return;
-  const normalized = normalizeProjectPath(projectName);
-  if (!normalized) return;
-  const hasOption = Array.from(refs.projectSelect.options).some(
-    (option) => normalizeProjectPath(option.value) === normalized,
-  );
-  if (hasOption) return;
-  const option = document.createElement("option");
-  option.value = normalized;
-  option.textContent = normalized;
-  refs.projectSelect.append(option);
-}
-
-function normalizeOnCreateAssistEnvelope(rawEnvelope) {
-  const suggestionsRaw = Array.isArray(rawEnvelope?.suggestions)
-    ? rawEnvelope.suggestions
-    : [];
-  const normalizedSuggestions = suggestionsRaw
-    .map((suggestion) => normalizeOnCreateSuggestion(suggestion))
-    .filter(Boolean)
-    .map((suggestion) => ({
-      ...suggestion,
-      rationale: truncateRationale(suggestion.rationale, 120),
-    }));
-  const sortedSuggestions = sortSuggestions(
-    ON_CREATE_SURFACE,
-    normalizedSuggestions,
-  );
-  const cappedSuggestions = capSuggestions(sortedSuggestions, 6);
-
-  let seenClarification = false;
-  for (const suggestion of cappedSuggestions) {
-    if (suggestion.type !== "ask_clarification") continue;
-    if (!seenClarification) {
-      seenClarification = true;
-      continue;
-    }
-    suggestion.dismissed = true;
-  }
-
-  return {
-    contractVersion: rawEnvelope?.contractVersion || 1,
-    generatedAt: String(rawEnvelope?.generatedAt || new Date().toISOString()),
-    requestId: String(rawEnvelope?.requestId || "on-create"),
-    surface: ON_CREATE_SURFACE,
-    must_abstain: !!rawEnvelope?.must_abstain,
-    suggestions: cappedSuggestions,
-  };
-}
-
-async function fetchOnCreateLatestSuggestion(todoId) {
-  return apiCall(
-    `${API_URL}/ai/suggestions/latest?todoId=${encodeURIComponent(todoId)}&surface=${ON_CREATE_SURFACE}`,
-  );
-}
-
-async function generateOnCreateSuggestion(todo) {
-  return apiCall(`${API_URL}/ai/decision-assist/stub`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({
-      surface: ON_CREATE_SURFACE,
-      todoId: todo.id,
-      title: todo.title,
-      description: todo.description || "",
-      notes: todo.notes || "",
-    }),
-  });
-}
-
-async function loadOnCreateDecisionAssist(todo, allowGenerate = true) {
-  if (!todo?.id) return;
-  const todoId = String(todo.id);
-  state.onCreateAssistState.loading = true;
-  state.onCreateAssistState.error = "";
-  state.onCreateAssistState.unavailable = false;
-  state.onCreateAssistState.mode = "live";
-  state.onCreateAssistState.liveTodoId = todoId;
-  state.onCreateAssistState.aiSuggestionId = "";
-  state.onCreateAssistState.envelope = null;
-  state.onCreateAssistState.suggestions = [];
-  state.onCreateAssistState.showAll = false;
-  renderOnCreateAssistRow();
-
-  try {
-    let latestResponse = await fetchOnCreateLatestSuggestion(todoId);
-    if (latestResponse.status === 403 || latestResponse.status === 404) {
-      state.onCreateAssistState.loading = false;
-      state.onCreateAssistState.unavailable = true;
-      renderOnCreateAssistRow();
-      return;
-    }
-
-    if (latestResponse.status === 204) {
-      if (isOnCreateDismissed(todoId) || !allowGenerate) {
-        state.onCreateAssistState.loading = false;
-        state.onCreateAssistState.envelope = normalizeOnCreateAssistEnvelope({
-          surface: ON_CREATE_SURFACE,
-          must_abstain: true,
-          suggestions: [],
-        });
-        state.onCreateAssistState.suggestions = [];
-        renderOnCreateAssistRow();
-        return;
-      }
-
-      const generated = await generateOnCreateSuggestion(todo);
-      if (generated.status === 403 || generated.status === 404) {
-        state.onCreateAssistState.loading = false;
-        state.onCreateAssistState.unavailable = true;
-        renderOnCreateAssistRow();
-        return;
-      }
-      latestResponse = await fetchOnCreateLatestSuggestion(todoId);
-    }
-
-    if (latestResponse.status === 204) {
-      state.onCreateAssistState.loading = false;
-      state.onCreateAssistState.envelope = normalizeOnCreateAssistEnvelope({
-        surface: ON_CREATE_SURFACE,
-        must_abstain: true,
-        suggestions: [],
-      });
-      state.onCreateAssistState.suggestions = [];
-      renderOnCreateAssistRow();
-      return;
-    }
-
-    if (!latestResponse.ok) {
-      state.onCreateAssistState.loading = false;
-      state.onCreateAssistState.error = "Could not load suggestions.";
-      renderOnCreateAssistRow();
-      return;
-    }
-
-    const payload = await latestResponse.json();
-    const envelope = normalizeOnCreateAssistEnvelope(
-      payload?.outputEnvelope || {},
-    );
-    state.onCreateAssistState.loading = false;
-    state.onCreateAssistState.aiSuggestionId = String(
-      payload?.aiSuggestionId || "",
-    );
-    state.onCreateAssistState.envelope = envelope;
-    state.onCreateAssistState.suggestions = envelope.suggestions;
-    state.onCreateAssistState.mode = "live";
-    state.onCreateAssistState.liveTodoId = todoId;
-    renderOnCreateAssistRow();
-  } catch (error) {
-    console.error("On-create AI load failed:", error);
-    state.onCreateAssistState.loading = false;
-    state.onCreateAssistState.error = "Could not load suggestions.";
-    renderOnCreateAssistRow();
-  }
-}
-
-function refreshOnCreateAssistFromTitle(force = false) {
-  const refs = getOnCreateAssistElements();
-  if (!refs) return;
-  const title = refs.titleInput.value.trim();
-  if (!title) {
-    resetOnCreateAssistState();
-    renderOnCreateAssistRow();
-    return;
-  }
-  if (!force && state.onCreateAssistState.titleBasis === title) {
-    renderOnCreateAssistRow();
-    return;
-  }
-  const envelope = normalizeOnCreateAssistEnvelope(
-    buildMockOnCreateAssistEnvelope(title),
-  );
-  state.onCreateAssistState = {
-    ...createInitialOnCreateAssistState(),
-    dismissedTodoIds: state.onCreateAssistState.dismissedTodoIds,
-    titleBasis: title,
-    envelope,
-    suggestions: envelope.suggestions,
-    mode: "mock",
-    showAll: false,
-  };
-  renderOnCreateAssistRow();
-}
-
-function getOnCreateSuggestionById(suggestionId) {
-  return state.onCreateAssistState.suggestions.find(
-    (suggestion) => suggestion.suggestionId === suggestionId,
-  );
-}
-
-function getActiveOnCreateSuggestions() {
-  return state.onCreateAssistState.suggestions.filter(
-    (suggestion) => !suggestion.dismissed,
-  );
-}
-
-function formatOnCreateDueDateLabel(dueDateIso) {
-  const date = new Date(dueDateIso);
-  if (Number.isNaN(date.getTime())) return "Set due date";
-  return `Due ${date.toLocaleDateString()}`;
-}
-
-function buildOnCreateChipSummary(suggestion) {
-  if (suggestion.type === "set_due_date") {
-    return formatOnCreateDueDateLabel(suggestion.payload.dueDateISO);
-  }
-  if (suggestion.type === "set_priority") {
-    return `Priority ${String(suggestion.payload.priority || "").toUpperCase() || "MEDIUM"}`;
-  }
-  if (suggestion.type === "set_project") {
-    return `Project ${String(suggestion.payload.projectName || suggestion.payload.category || "suggested")}`;
-  }
-  if (suggestion.type === "set_category") {
-    return `Category ${String(suggestion.payload.category || "suggested")}`;
-  }
-  if (suggestion.type === "rewrite_title") {
-    return "Refine title";
-  }
-  if (suggestion.type === "ask_clarification") {
-    return "Need one choice";
-  }
-  return "Suggestion";
-}
-
-function renderOnCreateChipChoices(suggestion) {
-  if (
-    suggestion.type !== "ask_clarification" ||
-    !suggestion.clarificationExpanded ||
-    suggestion.clarificationAnswered
-  ) {
-    return "";
-  }
-  const choices = Array.isArray(suggestion.payload.choices)
-    ? suggestion.payload.choices
-    : [];
-  if (choices.length === 0) return "";
-  return `
-    <div
-      class="ai-create-chip__choices"
-      role="radiogroup"
-      aria-label="Clarification choices"
-      aria-describedby="ai-create-rationale-${escapeHtml(suggestion.suggestionId)}"
-    >
-      ${choices
-        .map(
-          (choice) => `
-            <button
-              type="button"
-              class="ai-create-chip__choice ai-action-btn"
-              role="radio"
-              aria-checked="false"
-              data-ai-create-action="choose"
-              data-ai-create-suggestion-id="${escapeHtml(suggestion.suggestionId)}"
-              data-ai-create-choice-value="${escapeHtml(String(choice.value || ""))}"
-              aria-label="Choose ${escapeHtml(String(choice.label || choice.value || ""))}"
-            >
-              ${escapeHtml(String(choice.label || choice.value || ""))}
-            </button>
-          `,
-        )
-        .join("")}
-    </div>
-  `;
-}
-
-function renderOnCreateChipActions(suggestion) {
-  const suggestionId = escapeHtml(suggestion.suggestionId);
-  const label = escapeHtml(formatOnCreateSuggestionLabel(suggestion.type));
-  const rationaleId = `ai-create-rationale-${suggestionId}`;
-  if (suggestion.applied) {
-    return `
-      <span class="ai-create-chip__applied">Applied</span>
-      <button
-        type="button"
-        class="ai-create-chip__undo ai-undo"
-        data-testid="ai-chip-undo-${suggestionId}"
-        aria-label="Undo ${label}"
-        aria-describedby="${rationaleId}"
-        data-ai-create-action="undo"
-        data-ai-create-suggestion-id="${suggestionId}"
-      >
-        Undo
-      </button>
-    `;
-  }
-  if (suggestion.requiresConfirmation && suggestion.confirmationOpen) {
-    return `
-      <div class="ai-confirm" role="group" aria-label="Confirm apply ${label}">
-        <button
-          type="button"
-          class="ai-create-chip__confirm ai-action-btn"
-          data-testid="ai-chip-confirm-${suggestionId}"
-          aria-label="Confirm apply ${label}"
-          aria-describedby="${rationaleId}"
-          data-ai-create-action="confirm-apply"
-          data-ai-create-suggestion-id="${suggestionId}"
-        >
-          Confirm
-        </button>
-        <button
-          type="button"
-          class="ai-create-chip__action ai-action-btn"
-          aria-label="Cancel confirmation for ${label}"
-          aria-describedby="${rationaleId}"
-          data-ai-create-action="cancel-confirm"
-          data-ai-create-suggestion-id="${suggestionId}"
-        >
-          Cancel
-        </button>
-      </div>
-    `;
-  }
-  if (
-    suggestion.type === "ask_clarification" &&
-    !suggestion.clarificationAnswered
-  ) {
-    return `
-      <button
-        type="button"
-        class="ai-create-chip__action ai-action-btn"
-        data-testid="ai-chip-apply-${suggestionId}"
-        aria-label="Open choices for ${label}"
-        aria-describedby="${rationaleId}"
-        data-ai-create-action="toggle-choices"
-        data-ai-create-suggestion-id="${suggestionId}"
-      >
-        Choose
-      </button>
-      <button
-        type="button"
-        class="ai-create-chip__dismiss ai-action-btn"
-        data-testid="ai-chip-dismiss-${suggestionId}"
-        aria-label="Dismiss ${label}"
-        aria-describedby="${rationaleId}"
-        data-ai-create-action="dismiss"
-        data-ai-create-suggestion-id="${suggestionId}"
-      >
-        ×
-      </button>
-    `;
-  }
-  return `
-    <button
-      type="button"
-      class="ai-create-chip__action ai-action-btn"
-      data-testid="ai-chip-apply-${suggestionId}"
-      aria-label="Apply ${label}"
-      aria-describedby="${rationaleId}"
-      data-ai-create-action="apply"
-      data-ai-create-suggestion-id="${suggestionId}"
-    >
-      Apply
-    </button>
-    <button
-      type="button"
-      class="ai-create-chip__dismiss ai-action-btn"
-      data-testid="ai-chip-dismiss-${suggestionId}"
-      aria-label="Dismiss ${label}"
-      aria-describedby="${rationaleId}"
-      data-ai-create-action="dismiss"
-      data-ai-create-suggestion-id="${suggestionId}"
-    >
-      ×
-    </button>
-  `;
-}
-
-function renderOnCreateAssistRow() {
-  const refs = getOnCreateAssistElements();
-  if (!refs) return;
-  const title = refs.titleInput.value.trim();
-  const hasLiveContext = !!state.onCreateAssistState.liveTodoId;
-  if (!title && !hasLiveContext && !state.onCreateAssistState.loading) {
-    refs.row.hidden = true;
-    refs.row.innerHTML = "";
-    return;
-  }
-
-  // Lint-first gate: show one chip only; full panel revealed by Fix/Review.
-  // In debug mode, always show the full panel so developers see all metadata.
-  if (!state.onCreateAssistState.showFullAssist && !AI_DEBUG_ENABLED) {
-    const issue = lintTodoFields({
-      title,
-      dueDate: refs.dueDateInput ? refs.dueDateInput.value : "",
-      priority: state.currentPriority,
-      allTodos: state.todos,
-      surface: "on_create",
-    });
-    state.onCreateAssistState.lintIssue = issue;
-    if (!issue) {
-      refs.row.hidden = true;
-      refs.row.innerHTML = "";
-    } else {
-      refs.row.hidden = false;
-      refs.row.innerHTML = renderLintChip(issue);
-    }
-    return;
-  }
-
-  refs.row.hidden = false;
-  if (state.onCreateAssistState.loading) {
-    refs.row.innerHTML = `
-      <div class="ai-create-assist__header">
-        <span class="ai-create-assist__title">AI Assist</span>
-      </div>
-      <div class="ai-create-assist__empty ai-empty" role="status">Loading suggestions...</div>
-    `;
-    return;
-  }
-  if (state.onCreateAssistState.unavailable) {
-    refs.row.innerHTML = `
-      <div class="ai-create-assist__header">
-        <span class="ai-create-assist__title">AI Assist</span>
-      </div>
-      <div class="ai-create-assist__empty ai-empty" role="status">AI Suggestions unavailable.</div>
-    `;
-    return;
-  }
-  if (state.onCreateAssistState.error) {
-    refs.row.innerHTML = `
-      <div class="ai-create-assist__header">
-        <span class="ai-create-assist__title">AI Assist</span>
-      </div>
-      <div class="ai-create-assist__empty ai-empty" role="status">No suggestions right now.</div>
-    `;
-    return;
-  }
-  const activeSuggestions = getActiveOnCreateSuggestions();
-  if (activeSuggestions.length === 0) {
-    refs.row.innerHTML = `
-      <div class="ai-create-assist__header">
-        <span class="ai-create-assist__title">AI Assist</span>
-      </div>
-      ${renderAiDebugMeta(state.onCreateAssistState.envelope || {})}
-      <div class="ai-create-assist__empty ai-empty" role="status">No suggestions right now.</div>
-    `;
-    return;
-  }
-
-  const defaultLimit = 4;
-  const visibleLimit = state.onCreateAssistState.showAll ? 6 : defaultLimit;
-  const visibleSuggestions = activeSuggestions.slice(0, visibleLimit);
-  const hiddenCount = Math.max(
-    0,
-    activeSuggestions.length - visibleSuggestions.length,
-  );
-
-  refs.row.innerHTML = `
-    <div class="ai-create-assist__header">
-      <span class="ai-create-assist__title">AI Assist</span>
-      ${
-        hiddenCount > 0
-          ? `
-          <button
-            type="button"
-            class="ai-create-assist__expand"
-            data-testid="ai-chip-expand-more"
-            data-ai-create-action="toggle-more"
-            aria-label="${state.onCreateAssistState.showAll ? "Show fewer suggestions" : `Show ${hiddenCount} more suggestions`}"
-          >
-            ${state.onCreateAssistState.showAll ? "Show less" : `+${hiddenCount} more`}
-          </button>
-        `
-          : ""
-      }
-    </div>
-    ${renderAiDebugMeta(state.onCreateAssistState.envelope || {})}
-    <div class="ai-create-assist__chips">
-      ${visibleSuggestions
-        .map((suggestion) => {
-          const confidenceLabel = getOnCreateConfidenceBadge(
-            suggestion.confidence,
-          );
-          return `
-            <div
-              class="ai-create-chip ai-card ${suggestion.applied ? "ai-create-chip--applied" : ""}"
-              data-testid="ai-chip-${escapeHtml(suggestion.suggestionId)}"
-            >
-              <div class="ai-create-chip__top">
-                <span class="ai-create-chip__label">${escapeHtml(formatOnCreateSuggestionLabel(suggestion.type))}</span>
-                <span
-                  class="ai-create-chip__confidence ai-badge ai-badge--${escapeHtml(confidenceBand(suggestion.confidence))}"
-                  aria-label="Confidence ${escapeHtml(confidenceLabel)}"
-                >
-                  ${escapeHtml(confidenceLabel)}
-                </span>
-              </div>
-              <div
-                class="ai-create-chip__summary"
-                id="ai-create-rationale-${escapeHtml(suggestion.suggestionId)}"
-                title="${escapeHtml(suggestion.rationale || buildOnCreateChipSummary(suggestion))}"
-              >
-                ${escapeHtml(buildOnCreateChipSummary(suggestion))}
-              </div>
-              <div class="ai-create-chip__rationale ai-tooltip">${escapeHtml(suggestion.rationale || buildOnCreateChipSummary(suggestion))}</div>
-              ${renderAiDebugSuggestionId(suggestion.suggestionId)}
-              ${
-                suggestion.clarificationAnswered
-                  ? `<div class="ai-create-chip__helper">Answer: ${escapeHtml(suggestion.clarificationAnswer)}</div>`
-                  : ""
-              }
-              ${suggestion.helperText ? `<div class="ai-create-chip__helper">${escapeHtml(suggestion.helperText)}</div>` : ""}
-              ${renderOnCreateChipChoices(suggestion)}
-              <div class="ai-create-chip__actions ai-actions">
-                ${renderOnCreateChipActions(suggestion)}
-              </div>
-            </div>
-          `;
-        })
-        .join("")}
-    </div>
-  `;
-
-  const confirmationSuggestion = visibleSuggestions.find(
-    (suggestion) =>
-      needsConfirmation(suggestion) && suggestion.confirmationOpen,
-  );
-  if (confirmationSuggestion) {
-    window.requestAnimationFrame(() => {
-      const confirmButton = refs.row.querySelector(
-        `[data-testid="ai-chip-confirm-${escapeSelectorValue(confirmationSuggestion.suggestionId)}"]`,
-      );
-      if (confirmButton instanceof HTMLElement) {
-        confirmButton.focus({ preventScroll: true });
-      }
-    });
-  }
-}
-
-function snapshotOnCreateDraftState() {
-  const refs = getOnCreateAssistElements();
-  if (!refs) return null;
-  return {
-    title: refs.titleInput.value,
-    dueDate:
-      refs.dueDateInput instanceof HTMLInputElement
-        ? refs.dueDateInput.value
-        : "",
-    project:
-      refs.projectSelect instanceof HTMLSelectElement
-        ? refs.projectSelect.value
-        : "",
-    priority: state.currentPriority,
-  };
-}
-
-function restoreOnCreateDraftState(snapshot) {
-  if (!snapshot) return;
-  const refs = getOnCreateAssistElements();
-  if (!refs) return;
-  state.suppressOnCreateAssistInput = true;
-  refs.titleInput.value = snapshot.title || "";
-  if (refs.dueDateInput instanceof HTMLInputElement) {
-    refs.dueDateInput.value = snapshot.dueDate || "";
-  }
-  if (refs.projectSelect instanceof HTMLSelectElement) {
-    if (snapshot.project) {
-      ensureOnCreateProjectOption(snapshot.project);
-    }
-    refs.projectSelect.value = snapshot.project || "";
-  }
-  if (
-    snapshot.priority === "low" ||
-    snapshot.priority === "medium" ||
-    snapshot.priority === "high"
-  ) {
-    setPriority(snapshot.priority);
-  }
-  state.suppressOnCreateAssistInput = false;
-}
-
-function applyOnCreateSuggestion(suggestion, clarificationChoice = "") {
-  const refs = getOnCreateAssistElements();
-  if (!refs) return;
-  suggestion.undoSnapshot = snapshotOnCreateDraftState();
-  suggestion.applied = true;
-  suggestion.confirmationOpen = false;
-  suggestion.helperText = "";
-
-  if (suggestion.type === "rewrite_title") {
-    const nextTitle = String(suggestion.payload.title || "").trim();
-    if (nextTitle) {
-      state.suppressOnCreateAssistInput = true;
-      refs.titleInput.value = nextTitle;
-      state.suppressOnCreateAssistInput = false;
-    }
-  } else if (suggestion.type === "set_due_date") {
-    const dueDateIso = String(suggestion.payload.dueDateISO || "");
-    if (refs.dueDateInput instanceof HTMLInputElement) {
-      refs.dueDateInput.value = toDateTimeLocalValue(dueDateIso);
-    }
-  } else if (suggestion.type === "set_priority") {
-    const nextPriority = String(
-      suggestion.payload.priority || "",
-    ).toLowerCase();
-    if (
-      nextPriority === "low" ||
-      nextPriority === "medium" ||
-      nextPriority === "high"
-    ) {
-      setPriority(nextPriority);
-    }
-  } else if (suggestion.type === "set_project") {
-    const rawProject = String(
-      suggestion.payload.projectName || suggestion.payload.category || "",
-    );
-    const normalized = normalizeProjectPath(rawProject);
-    if (normalized && refs.projectSelect instanceof HTMLSelectElement) {
-      ensureOnCreateProjectOption(normalized);
-      refs.projectSelect.value = normalized;
-    }
-  } else if (suggestion.type === "set_category") {
-    const rawCategory = String(suggestion.payload.category || "");
-    const normalized = normalizeProjectPath(rawCategory);
-    if (normalized && refs.projectSelect instanceof HTMLSelectElement) {
-      ensureOnCreateProjectOption(normalized);
-      refs.projectSelect.value = normalized;
-    }
-  } else if (suggestion.type === "ask_clarification") {
-    suggestion.clarificationAnswered = true;
-    suggestion.clarificationExpanded = false;
-    suggestion.clarificationAnswer = clarificationChoice || "Selected";
-    const selectedChoice =
-      (Array.isArray(suggestion.payload.choices)
-        ? suggestion.payload.choices.find(
-            (choice) => String(choice.value || "") === clarificationChoice,
-          )
-        : null) || null;
-    const projectValue = normalizeProjectPath(
-      String(
-        selectedChoice?.projectName ||
-          selectedChoice?.category ||
-          selectedChoice?.value ||
-          "",
-      ),
-    );
-    if (projectValue && refs.projectSelect instanceof HTMLSelectElement) {
-      ensureOnCreateProjectOption(projectValue);
-      refs.projectSelect.value = projectValue;
-    }
-    suggestion.helperText = "Thanks — will refine next time.";
-  }
-}
-
-async function applyLiveOnCreateSuggestion(suggestion, confirmed = false) {
-  if (!state.onCreateAssistState.aiSuggestionId) return;
-  const response = await apiCall(
-    `${API_URL}/ai/suggestions/${encodeURIComponent(state.onCreateAssistState.aiSuggestionId)}/apply`,
-    {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        suggestionId: suggestion.suggestionId,
-        confirmed: confirmed === true,
-      }),
-    },
-  );
-  if (!response.ok) {
-    const data = await response.json().catch(() => ({}));
-    state.onCreateAssistState.error =
-      typeof data?.error === "string"
-        ? data.error
-        : "Could not apply suggestion.";
-    renderOnCreateAssistRow();
-    return;
-  }
-
-  const data = await response.json().catch(() => ({}));
-  if (data?.todo?.id) {
-    const index = state.todos.findIndex((item) => item.id === data.todo.id);
-    if (index >= 0) {
-      state.todos[index] = data.todo;
-    }
-    renderTodos();
-  }
-  clearOnCreateDismissed(state.onCreateAssistState.liveTodoId);
-  const refreshedTodo =
-    (state.onCreateAssistState.liveTodoId &&
-      state.todos.find(
-        (item) => item.id === state.onCreateAssistState.liveTodoId,
-      )) ||
-    null;
-  if (refreshedTodo) {
-    await loadOnCreateDecisionAssist(refreshedTodo, false);
-  } else {
-    resetOnCreateAssistState();
-    renderOnCreateAssistRow();
-  }
-}
-
-async function onCreateAssistApplySuggestion(suggestionId) {
-  const suggestion = getOnCreateSuggestionById(suggestionId);
-  if (!suggestion || suggestion.dismissed || suggestion.applied) return;
-  if (suggestion.type === "ask_clarification") {
-    suggestion.clarificationExpanded = !suggestion.clarificationExpanded;
-    renderOnCreateAssistRow();
-    return;
-  }
-  if (suggestion.requiresConfirmation) {
-    suggestion.confirmationOpen = true;
-    renderOnCreateAssistRow();
-    return;
-  }
-  if (state.onCreateAssistState.mode === "live") {
-    await applyLiveOnCreateSuggestion(suggestion, false);
-    return;
-  }
-  applyOnCreateSuggestion(suggestion);
-  renderOnCreateAssistRow();
-}
-
-async function onCreateAssistConfirmApplySuggestion(suggestionId) {
-  const suggestion = getOnCreateSuggestionById(suggestionId);
-  if (!suggestion || suggestion.dismissed || suggestion.applied) return;
-  if (state.onCreateAssistState.mode === "live") {
-    await applyLiveOnCreateSuggestion(suggestion, true);
-    return;
-  }
-  applyOnCreateSuggestion(suggestion);
-  renderOnCreateAssistRow();
-}
-
-async function onCreateAssistDismissSuggestion(suggestionId) {
-  const suggestion = getOnCreateSuggestionById(suggestionId);
-  if (!suggestion) return;
-  if (
-    state.onCreateAssistState.mode === "live" &&
-    state.onCreateAssistState.aiSuggestionId
-  ) {
-    try {
-      await apiCall(
-        `${API_URL}/ai/suggestions/${encodeURIComponent(state.onCreateAssistState.aiSuggestionId)}/dismiss`,
-        {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ suggestionId, dismissAll: true }),
-        },
-      );
-    } catch (error) {
-      console.error("On-create AI dismiss failed:", error);
-    }
-    markOnCreateDismissed(state.onCreateAssistState.liveTodoId);
-    state.onCreateAssistState.suggestions = [];
-    state.onCreateAssistState.envelope = normalizeOnCreateAssistEnvelope({
-      surface: ON_CREATE_SURFACE,
-      must_abstain: true,
-      suggestions: [],
-    });
-    state.onCreateAssistState.aiSuggestionId = "";
-    state.onCreateAssistState.error = "";
-    renderOnCreateAssistRow();
-    return;
-  }
-  suggestion.dismissed = true;
-  renderOnCreateAssistRow();
-}
-
-function onCreateAssistUndoSuggestion(suggestionId) {
-  const suggestion = getOnCreateSuggestionById(suggestionId);
-  if (!suggestion || !suggestion.applied) return;
-  restoreOnCreateDraftState(suggestion.undoSnapshot);
-  suggestion.applied = false;
-  suggestion.confirmationOpen = false;
-  suggestion.clarificationAnswered = false;
-  suggestion.clarificationAnswer = "";
-  suggestion.helperText = "";
-  suggestion.clarificationExpanded = false;
-  suggestion.undoSnapshot = null;
-  emitAiSuggestionUndoTelemetry({
-    surface: ON_CREATE_SURFACE,
-    aiSuggestionDbId: state.onCreateAssistState.aiSuggestionId,
-    suggestionId,
-    todoId: state.onCreateAssistState.liveTodoId || "",
-    selectedTodoIdsCount: 1,
-  });
-  renderOnCreateAssistRow();
-}
-
-function onCreateAssistChooseClarification(suggestionId, choiceValue) {
-  const suggestion = getOnCreateSuggestionById(suggestionId);
-  if (
-    !suggestion ||
-    suggestion.type !== "ask_clarification" ||
-    suggestion.applied
-  ) {
-    return;
-  }
-  applyOnCreateSuggestion(suggestion, String(choiceValue || "").trim());
-  renderOnCreateAssistRow();
-}
-
-function bindOnCreateAssistHandlers() {
-  if (window.__onCreateAssistHandlersBound) {
-    return;
-  }
-  window.__onCreateAssistHandlersBound = true;
-
-  document.addEventListener("input", (event) => {
-    const target = event.target;
-    if (!(target instanceof HTMLElement)) return;
-    if (target.id !== "todoInput") return;
-    if (state.suppressOnCreateAssistInput) {
-      renderOnCreateAssistRow();
-      return;
-    }
-    refreshOnCreateAssistFromTitle();
-  });
-
-  // Lint chip Fix/Review delegation (on-create surface only).
-  // Excludes clicks inside #todoDetailsDrawer which has its own handler.
-  document.addEventListener("click", async (event) => {
-    const target = event.target;
-    if (!(target instanceof Element)) return;
-    if (target.closest("#todoDetailsDrawer")) return;
-    const lintEl = target.closest("[data-ai-lint-action]");
-    if (!(lintEl instanceof HTMLElement)) return;
-    const lintAction = lintEl.getAttribute("data-ai-lint-action");
-    state.onCreateAssistState.showFullAssist = true;
-    if (state.onCreateAssistState.liveTodoId) {
-      // Post-create mode: load server-backed suggestions for the saved todo.
-      const todo = getTodoById(state.onCreateAssistState.liveTodoId);
-      await loadOnCreateDecisionAssist(todo, lintAction === "fix");
-    } else {
-      // Pre-create mode: regenerate client-side mock suggestions and reveal them.
-      // refreshOnCreateAssistFromTitle resets state; set showFullAssist after.
-      refreshOnCreateAssistFromTitle(true);
-      state.onCreateAssistState.showFullAssist = true;
-      renderOnCreateAssistRow();
-    }
-  });
-
-  document.addEventListener("click", async (event) => {
-    const target = event.target;
-    if (!(target instanceof Element)) return;
-    const actionEl = target.closest("[data-ai-create-action]");
-    if (!(actionEl instanceof HTMLElement)) return;
-    const action = actionEl.getAttribute("data-ai-create-action");
-    const suggestionId =
-      actionEl.getAttribute("data-ai-create-suggestion-id") || "";
-    if (action === "toggle-more") {
-      state.onCreateAssistState.showAll = !state.onCreateAssistState.showAll;
-      renderOnCreateAssistRow();
-      return;
-    }
-    if (action === "apply" || action === "toggle-choices") {
-      await onCreateAssistApplySuggestion(suggestionId);
-      return;
-    }
-    if (action === "confirm-apply") {
-      await onCreateAssistConfirmApplySuggestion(suggestionId);
-      return;
-    }
-    if (action === "cancel-confirm") {
-      const suggestion = getOnCreateSuggestionById(suggestionId);
-      if (!suggestion) return;
-      suggestion.confirmationOpen = false;
-      renderOnCreateAssistRow();
-      return;
-    }
-    if (action === "dismiss") {
-      await onCreateAssistDismissSuggestion(suggestionId);
-      return;
-    }
-    if (action === "undo") {
-      onCreateAssistUndoSuggestion(suggestionId);
-      return;
-    }
-    if (action === "choose") {
-      const choiceValue =
-        actionEl.getAttribute("data-ai-create-choice-value") || "";
-      onCreateAssistChooseClarification(suggestionId, choiceValue);
-      return;
-    }
-  });
-
-  document.addEventListener("keydown", (event) => {
-    if (event.key !== "Escape") return;
-    if (
-      !state.onCreateAssistState.suggestions.some(
-        (item) => item.confirmationOpen,
-      )
-    ) {
-      return;
-    }
-    const target = event.target;
-    if (target instanceof Element && !target.closest("#aiOnCreateAssistRow")) {
-      return;
-    }
-    state.onCreateAssistState.suggestions.forEach((item) => {
-      item.confirmationOpen = false;
-    });
-    renderOnCreateAssistRow();
-  });
-}
-
-function resetTodayPlanState() {
-  state.todayPlanState = createInitialTodayPlanState();
-}
-
-function getTodayPlanPanelElement() {
-  const panel = document.getElementById("todayPlanPanel");
-  if (!(panel instanceof HTMLElement)) return null;
-  return panel;
-}
-
-function getTodayPlanImpactRank(type) {
-  return impactRankForSurface(TODAY_PLAN_SURFACE, type);
-}
-
-function getTodayPlanConfidenceBadge(confidence) {
-  return confidenceLabel(confidence);
-}
-
-function isTodayPlanViewActive() {
-  return state.currentDateView === "today";
-}
-
-function toEpoch(value) {
-  const date = value ? new Date(value) : null;
-  if (!date || Number.isNaN(date.getTime())) return Infinity;
-  return date.getTime();
-}
-
-function normalizePriorityValue(priority) {
-  const value = String(priority || "").toLowerCase();
-  if (value === "low" || value === "medium" || value === "high") {
-    return value;
-  }
-  return "medium";
-}
-
-function priorityWeight(priority) {
-  const value = normalizePriorityValue(priority);
-  if (value === "high") return 3;
-  if (value === "medium") return 2;
-  return 1;
-}
-
-function estimateTodoMinutes(todo, mode = "balanced") {
-  const title = String(todo?.title || "").trim();
-  const words = title ? title.split(/\s+/).length : 1;
-  const base = Math.max(15, Math.min(120, words * 9));
-  if (mode === "quick") {
-    return Math.max(10, Math.min(30, Math.round(base * 0.6)));
-  }
-  if (mode === "deep") {
-    return Math.max(45, Math.min(150, Math.round(base * 1.35)));
-  }
-  return base;
-}
-
-function rankTodayTodos(goalLower, todayTodos) {
-  const mode = goalLower.includes("quick")
-    ? "quick"
-    : goalLower.includes("deep") || goalLower.includes("focus")
-      ? "deep"
-      : "balanced";
-  const now = Date.now();
-  const ranked = [...todayTodos].sort((a, b) => {
-    const dueA = toEpoch(a.dueDate);
-    const dueB = toEpoch(b.dueDate);
-    const dueScoreA = dueA === Infinity ? Infinity : Math.max(0, dueA - now);
-    const dueScoreB = dueB === Infinity ? Infinity : Math.max(0, dueB - now);
-    const recencyA = toEpoch(a.updatedAt || a.createdAt);
-    const recencyB = toEpoch(b.updatedAt || b.createdAt);
-    const prioA = priorityWeight(a.priority);
-    const prioB = priorityWeight(b.priority);
-    if (mode === "quick") {
-      const quickA = estimateTodoMinutes(a, "quick");
-      const quickB = estimateTodoMinutes(b, "quick");
-      if (quickA !== quickB) return quickA - quickB;
-      if (dueScoreA !== dueScoreB) return dueScoreA - dueScoreB;
-      if (prioA !== prioB) return prioB - prioA;
-      return recencyB - recencyA;
-    }
-    if (mode === "deep") {
-      const deepA = estimateTodoMinutes(a, "deep");
-      const deepB = estimateTodoMinutes(b, "deep");
-      if (deepA !== deepB) return deepB - deepA;
-      if (prioA !== prioB) return prioB - prioA;
-      if (dueScoreA !== dueScoreB) return dueScoreA - dueScoreB;
-      return recencyB - recencyA;
-    }
-    if (dueScoreA !== dueScoreB) return dueScoreA - dueScoreB;
-    if (prioA !== prioB) return prioB - prioA;
-    return recencyB - recencyA;
-  });
-  return { mode, ranked };
-}
-
-function buildTodayPlanSuggestion(
-  type,
-  suggestionId,
-  confidence,
-  rationale,
-  payload,
-) {
-  return {
-    type,
-    suggestionId,
-    confidence,
-    rationale: truncateRationale(rationale, 120),
-    payload,
-  };
-}
-
-function mockPlanFromGoal(goalText, todayTodos) {
-  const goal = String(goalText || "").trim();
-  const goalLower = goal.toLowerCase();
-  const generatedAt = new Date().toISOString();
-
-  if (
-    goalLower.includes("abstain") ||
-    !Array.isArray(todayTodos) ||
-    !todayTodos.length
-  ) {
-    return {
-      contractVersion: 1,
-      generatedAt,
-      requestId: `today-plan-${encodeURIComponent(goalLower || "empty")}`,
-      surface: TODAY_PLAN_SURFACE,
-      must_abstain: true,
-      planPreview: { topN: 0, items: [] },
-      suggestions: [],
-    };
-  }
-
-  const { mode, ranked } = rankTodayTodos(goalLower, todayTodos);
-  const topN =
-    ranked.length >= 5 && (mode === "quick" || mode === "deep") ? 5 : 3;
-  const selected = ranked.slice(0, Math.min(topN, ranked.length));
-  const previewItems = selected.map((todo, index) => {
-    const estimateMode =
-      mode === "quick"
-        ? "quick"
-        : mode === "deep" && index === 0
-          ? "deep"
-          : "balanced";
-    const minutes = estimateTodoMinutes(todo, estimateMode);
-    const rationale =
-      mode === "quick"
-        ? "Quick-win candidate with low setup cost."
-        : mode === "deep" && index === 0
-          ? "Primary focus block for deep work."
-          : "Urgency and priority alignment for today.";
-    return {
-      todoId: String(todo.id),
-      rank: index + 1,
-      timeEstimateMin: minutes,
-      rationale,
-    };
-  });
-
-  const suggestions = [];
-  for (const item of previewItems) {
-    const todo = selected.find((entry) => String(entry.id) === item.todoId);
-    if (!todo) continue;
-    if (item.rank <= 2) {
-      const due = new Date();
-      due.setDate(due.getDate() + item.rank);
-      due.setHours(9 + item.rank, 0, 0, 0);
-      suggestions.push(
-        buildTodayPlanSuggestion(
-          "set_due_date",
-          `today-set-due-${item.todoId}`,
-          0.8 - item.rank * 0.04,
-          "Assign a concrete deadline for today's execution.",
-          { todoId: item.todoId, dueDateISO: due.toISOString() },
-        ),
-      );
-    }
-    if (normalizePriorityValue(todo.priority) !== "high" && item.rank === 1) {
-      suggestions.push(
-        buildTodayPlanSuggestion(
-          "set_priority",
-          `today-set-priority-${item.todoId}`,
-          0.77,
-          "Top ranked item should be elevated for focus.",
-          { todoId: item.todoId, priority: "high" },
-        ),
-      );
-    }
-    if (item.rank <= 2) {
-      suggestions.push(
-        buildTodayPlanSuggestion(
-          "propose_next_action",
-          `today-next-action-${item.todoId}`,
-          0.7,
-          "Define the first concrete step before context switching.",
-          {
-            todoId: item.todoId,
-            text: `Start: ${String(todo.title || "").slice(0, 80)} and draft first deliverable.`,
-          },
-        ),
-      );
-    }
-  }
-
-  if (previewItems[0]) {
-    const todoId = previewItems[0].todoId;
-    suggestions.push(
-      buildTodayPlanSuggestion(
-        "split_subtasks",
-        `today-split-${todoId}`,
-        0.68,
-        "Split one larger item into scoped execution steps.",
-        {
-          todoId,
-          subtasks: [
-            { title: "Outline deliverable", order: 1 },
-            { title: "Execute focused work block", order: 2 },
-            { title: "Review and close loop", order: 3 },
-          ],
-        },
-      ),
-    );
-  }
-
-  suggestions.push({
-    type: "unknown_type",
-    suggestionId: "today-unknown",
-    confidence: 0.5,
-    rationale: "Should be ignored",
-    payload: { todoId: previewItems[0]?.todoId || "" },
-  });
-
-  return {
-    contractVersion: 1,
-    generatedAt,
-    requestId: `today-plan-${encodeURIComponent(goalLower || "none")}-${todayTodos.length}`,
-    surface: TODAY_PLAN_SURFACE,
-    must_abstain: false,
-    planPreview: {
-      topN: previewItems.length,
-      items: previewItems,
-    },
-    suggestions,
-  };
-}
-
-function normalizeTodayPlanEnvelope(rawEnvelope) {
-  const suggestionsRaw = Array.isArray(rawEnvelope?.suggestions)
-    ? rawEnvelope.suggestions
-    : [];
-  const normalizedSuggestions = suggestionsRaw
-    .filter((suggestion) =>
-      shouldRenderTypeForSurface(TODAY_PLAN_SURFACE, suggestion?.type),
-    )
-    .map((suggestion) => ({
-      type: String(suggestion.type),
-      suggestionId: String(suggestion.suggestionId || ""),
-      confidence: Math.max(0, Math.min(1, Number(suggestion.confidence) || 0)),
-      rationale: truncateRationale(suggestion.rationale, 120),
-      payload:
-        suggestion.payload && typeof suggestion.payload === "object"
-          ? suggestion.payload
-          : {},
-    }))
-    .filter((suggestion) => suggestion.suggestionId);
-
-  const previewItemsRaw = Array.isArray(rawEnvelope?.planPreview?.items)
-    ? rawEnvelope.planPreview.items
-    : [];
-  const previewItems = previewItemsRaw
-    .map((item) => ({
-      todoId: String(item?.todoId || ""),
-      rank: Number(item?.rank) || 0,
-      timeEstimateMin: Number(item?.timeEstimateMin) || 0,
-      rationale: truncateRationale(item?.rationale, 120),
-    }))
-    .filter((item) => item.todoId && item.rank > 0);
-
-  return {
-    contractVersion: Number(rawEnvelope?.contractVersion) || 1,
-    generatedAt: String(rawEnvelope?.generatedAt || new Date().toISOString()),
-    requestId: String(rawEnvelope?.requestId || "today-plan"),
-    surface: TODAY_PLAN_SURFACE,
-    must_abstain: !!rawEnvelope?.must_abstain,
-    planPreview: {
-      topN: Number(rawEnvelope?.planPreview?.topN) || previewItems.length,
-      items: previewItems,
-    },
-    suggestions: normalizedSuggestions,
-  };
-}
-
-async function fetchTodayPlanLatestSuggestion() {
-  return apiCall(
-    `${API_URL}/ai/suggestions/latest?surface=${TODAY_PLAN_SURFACE}`,
-  );
-}
-
-function buildTodayPlanCandidates() {
-  return getVisibleTodos().map((todo) => ({
-    id: String(todo.id),
-    title: String(todo.title || ""),
-    dueDate: todo.dueDate || undefined,
-    priority: todo.priority || undefined,
-    createdAt: todo.createdAt || undefined,
-    updatedAt: todo.updatedAt || undefined,
-  }));
-}
-
-async function generateTodayPlanSuggestion(goalText) {
-  const candidates = buildTodayPlanCandidates();
-  const preferredTopN = candidates.length >= 5 ? 5 : 3;
-  return apiCall(`${API_URL}/ai/decision-assist/stub`, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      "x-ai-explicit-request": "1",
-    },
-    body: JSON.stringify({
-      surface: TODAY_PLAN_SURFACE,
-      goal: goalText || undefined,
-      topN: preferredTopN,
-      timeZone: Intl.DateTimeFormat().resolvedOptions().timeZone || "UTC",
-      anchorDateISO: new Date().toISOString(),
-      todoCandidates: candidates,
-    }),
-  });
-}
-
-async function loadTodayPlanDecisionAssist(allowGenerate = false) {
-  if (!isTodayPlanViewActive()) return;
-  if (!FEATURE_TASK_DRAWER_DECISION_ASSIST) {
-    state.todayPlanState.loading = false;
-    state.todayPlanState.generating = false;
-    state.todayPlanState.unavailable = true;
-    state.todayPlanState.hasLoaded = true;
-    renderTodayPlanPanel();
-    return;
-  }
-
-  state.todayPlanState.loading = true;
-  state.todayPlanState.error = "";
-  state.todayPlanState.unavailable = false;
-  renderTodayPlanPanel();
-
-  try {
-    let latestResponse = await fetchTodayPlanLatestSuggestion();
-    if (latestResponse.status === 403 || latestResponse.status === 404) {
-      state.todayPlanState.loading = false;
-      state.todayPlanState.generating = false;
-      state.todayPlanState.unavailable = true;
-      state.todayPlanState.hasLoaded = true;
-      renderTodayPlanPanel();
-      return;
-    }
-
-    if (latestResponse.status === 204 && allowGenerate) {
-      const generated = await generateTodayPlanSuggestion(
-        state.todayPlanState.goalText,
-      );
-      if (generated.status === 403 || generated.status === 404) {
-        state.todayPlanState.loading = false;
-        state.todayPlanState.generating = false;
-        state.todayPlanState.unavailable = true;
-        state.todayPlanState.hasLoaded = true;
-        renderTodayPlanPanel();
-        return;
-      }
-      latestResponse = await fetchTodayPlanLatestSuggestion();
-    }
-
-    if (latestResponse.status === 204) {
-      state.todayPlanState.loading = false;
-      state.todayPlanState.generating = false;
-      state.todayPlanState.hasLoaded = true;
-      state.todayPlanState.aiSuggestionId = "";
-      state.todayPlanState.envelope = normalizeTodayPlanEnvelope({
-        surface: TODAY_PLAN_SURFACE,
-        must_abstain: false,
-        planPreview: { topN: 3, items: [] },
-        suggestions: [],
-      });
-      state.todayPlanState.selectedTodoIds = new Set();
-      state.todayPlanState.dismissedSuggestionIds = new Set();
-      renderTodayPlanPanel();
-      return;
-    }
-
-    if (!latestResponse.ok) {
-      state.todayPlanState.loading = false;
-      state.todayPlanState.generating = false;
-      state.todayPlanState.error = "Could not load suggestions.";
-      state.todayPlanState.hasLoaded = true;
-      renderTodayPlanPanel();
-      return;
-    }
-
-    const payload = await latestResponse.json();
-    const envelope = normalizeTodayPlanEnvelope(payload?.outputEnvelope || {});
-    state.todayPlanState.loading = false;
-    state.todayPlanState.generating = false;
-    state.todayPlanState.hasLoaded = true;
-    state.todayPlanState.aiSuggestionId = String(payload?.aiSuggestionId || "");
-    state.todayPlanState.envelope = envelope;
-    state.todayPlanState.dismissedSuggestionIds = new Set();
-    state.todayPlanState.selectedTodoIds = new Set(
-      envelope.planPreview.items
-        .map((item) => String(item.todoId))
-        .filter(Boolean),
-    );
-    renderTodayPlanPanel();
-  } catch (error) {
-    console.error("Today plan AI load failed:", error);
-    state.todayPlanState.loading = false;
-    state.todayPlanState.generating = false;
-    state.todayPlanState.error = "Could not load suggestions.";
-    state.todayPlanState.hasLoaded = true;
-    renderTodayPlanPanel();
-  }
-}
-
-function getTodayPlanSelectedSuggestionCards() {
-  if (!state.todayPlanState.envelope) return [];
-  const selectedIds = state.todayPlanState.selectedTodoIds;
-  const filtered = state.todayPlanState.envelope.suggestions
-    .filter((suggestion) => {
-      const todoId = String(suggestion.payload?.todoId || "");
-      return todoId && selectedIds.has(todoId);
-    })
-    .filter(
-      (suggestion) =>
-        !state.todayPlanState.dismissedSuggestionIds.has(
-          suggestion.suggestionId,
-        ),
-    )
-    .filter((suggestion) =>
-      shouldRenderTypeForSurface(TODAY_PLAN_SURFACE, suggestion.type),
-    );
-  return capSuggestions(sortSuggestions(TODAY_PLAN_SURFACE, filtered), 6);
-}
-
-function deepClone(value) {
-  return JSON.parse(JSON.stringify(value));
-}
-
-function renderTodayPlanPanel() {
-  const panel = getTodayPlanPanelElement();
-  if (!panel) return;
-  if (!isTodayPlanViewActive()) {
-    panel.hidden = true;
-    panel.innerHTML = "";
-    return;
-  }
-
-  panel.hidden = false;
-  const goalText = state.todayPlanState.goalText || "";
-  const envelope = state.todayPlanState.envelope;
-  const previewItems = Array.isArray(envelope?.planPreview?.items)
-    ? envelope.planPreview.items
-    : [];
-  const suggestionCards = getTodayPlanSelectedSuggestionCards();
-  const emptyMessage = envelope?.must_abstain
-    ? "No safe plan right now."
-    : envelope &&
-        !state.todayPlanState.generating &&
-        suggestionCards.length === 0
-      ? "No suggestions right now."
-      : "";
-
-  panel.setAttribute("data-testid", "today-plan-panel");
-  panel.innerHTML = `
-    <div class="today-plan-panel__header">
-      <div class="today-plan-panel__title">Plan my day</div>
-      ${
-        state.todayPlanState.lastApplyBatch
-          ? `
-          <button
-            type="button"
-            class="today-plan-panel__undo ai-undo"
-            data-testid="today-plan-undo"
-            aria-label="Undo last plan apply"
-            data-today-plan-action="undo"
-          >
-            Undo
-          </button>
-        `
-          : ""
-      }
-    </div>
-    <div class="today-plan-panel__controls">
-      <label class="sr-only" for="todayPlanGoalInput">Goal (optional)</label>
-      <input
-        id="todayPlanGoalInput"
-        data-testid="today-plan-goal-input"
-        type="text"
-        placeholder="Goal (optional)"
-        value="${escapeHtml(goalText)}"
-        aria-label="Goal (optional)"
-      />
-      <button
-        id="todayPlanGenerateButton"
-        data-testid="today-plan-generate"
-        type="button"
-        class="mini-btn"
-        aria-label="Generate plan"
-        data-today-plan-action="generate"
-        ${state.todayPlanState.loading || state.todayPlanState.generating ? "disabled" : ""}
-      >
-        ${state.todayPlanState.loading || state.todayPlanState.generating ? "Generating..." : "Generate plan"}
-      </button>
-    </div>
-    ${renderAiDebugMeta(envelope || {})}
-    ${
-      state.todayPlanState.loading || state.todayPlanState.generating
-        ? '<div class="today-plan-panel__loading ai-empty" role="status">Generating plan preview...</div>'
-        : ""
-    }
-    ${
-      state.todayPlanState.unavailable
-        ? '<div class="today-plan-panel__empty ai-empty" role="status">AI Suggestions unavailable.</div>'
-        : ""
-    }
-    ${
-      state.todayPlanState.error
-        ? `<div class="today-plan-panel__empty ai-empty" role="status">${escapeHtml(state.todayPlanState.error)}</div>`
-        : ""
-    }
-    ${
-      previewItems.length > 0
-        ? `
-        <div class="today-plan-preview" data-testid="today-plan-preview">
-          ${previewItems
-            .map((item) => {
-              const todo = state.todos.find(
-                (entry) => String(entry.id) === String(item.todoId),
-              );
-              const checked = state.todayPlanState.selectedTodoIds.has(
-                item.todoId,
-              );
-              return `
-                <div class="today-plan-preview__item" data-testid="today-plan-item-${escapeHtml(item.todoId)}">
-                  <input
-                    type="checkbox"
-                    data-testid="today-plan-item-checkbox-${escapeHtml(item.todoId)}"
-                    data-today-plan-action="toggle-item"
-                    data-today-plan-todo-id="${escapeHtml(item.todoId)}"
-                    aria-label="Select plan item ${escapeHtml(String(todo?.title || item.todoId))}"
-                    ${checked ? "checked" : ""}
-                  />
-                  <div class="today-plan-preview__rank">${item.rank}</div>
-                  <div class="today-plan-preview__body">
-                    <div class="today-plan-preview__title">${escapeHtml(String(todo?.title || "Task"))}</div>
-                    <div class="today-plan-preview__meta">${item.timeEstimateMin} min • ${escapeHtml(item.rationale)}</div>
-                  </div>
-                </div>
-              `;
-            })
-            .join("")}
-        </div>
-      `
-        : ""
-    }
-    ${emptyMessage ? `<div class="today-plan-panel__empty ai-empty" role="status">${escapeHtml(emptyMessage)}</div>` : ""}
-    ${
-      suggestionCards.length > 0
-        ? `
-        <div class="today-plan-suggestions">
-          ${suggestionCards
-            .map((suggestion) => {
-              const todoId = String(suggestion.payload?.todoId || "");
-              const todo = state.todos.find(
-                (entry) => String(entry.id) === todoId,
-              );
-              const summary =
-                suggestion.type === "set_due_date"
-                  ? `Set due ${new Date(String(suggestion.payload?.dueDateISO || "")).toLocaleDateString()}`
-                  : suggestion.type === "set_priority"
-                    ? `Set priority ${String(suggestion.payload?.priority || "").toUpperCase()}`
-                    : suggestion.type === "split_subtasks"
-                      ? `Split into ${Array.isArray(suggestion.payload?.subtasks) ? suggestion.payload.subtasks.length : 0} subtasks`
-                      : `Next action: ${String(suggestion.payload?.text || "").slice(0, 60)}`;
-              return `
-                <div
-                  class="today-plan-suggestion ai-card"
-                  data-testid="today-plan-suggestion-${escapeHtml(suggestion.suggestionId)}"
-                  data-today-plan-todo-id="${escapeHtml(todoId)}"
-                >
-                  <div class="today-plan-suggestion__title">${escapeHtml(labelForType(suggestion.type))}</div>
-                  <div class="today-plan-suggestion__summary">${escapeHtml(summary)}</div>
-                  <div
-                    class="today-plan-suggestion__rationale ai-tooltip"
-                    id="today-plan-rationale-${escapeHtml(suggestion.suggestionId)}"
-                  >
-                    ${escapeHtml(suggestion.rationale)}
-                  </div>
-                  ${renderAiDebugSuggestionId(suggestion.suggestionId)}
-                  <div class="today-plan-suggestion__footer ai-actions">
-                    <span class="today-plan-suggestion__confidence ai-badge ai-badge--${escapeHtml(confidenceBand(suggestion.confidence))}" aria-label="Confidence ${escapeHtml(getTodayPlanConfidenceBadge(suggestion.confidence))}">
-                      ${escapeHtml(getTodayPlanConfidenceBadge(suggestion.confidence))}
-                    </span>
-                    <button
-                      type="button"
-                      class="today-plan-suggestion__dismiss ai-action-btn"
-                      data-testid="today-plan-suggestion-dismiss-${escapeHtml(suggestion.suggestionId)}"
-                      aria-label="Dismiss suggestion ${escapeHtml(suggestion.suggestionId)}"
-                      aria-describedby="today-plan-rationale-${escapeHtml(suggestion.suggestionId)}"
-                      data-today-plan-action="dismiss-suggestion"
-                      data-today-plan-suggestion-id="${escapeHtml(suggestion.suggestionId)}"
-                    >
-                      Dismiss
-                    </button>
-                  </div>
-                </div>
-              `;
-            })
-            .join("")}
-        </div>
-        <div class="today-plan-panel__actions">
-          <button
-            type="button"
-            class="add-btn"
-            data-testid="today-plan-apply-selected"
-            aria-label="Apply selected plan suggestions"
-            data-today-plan-action="apply-selected"
-          >
-            Apply selected
-          </button>
-        </div>
-      `
-        : ""
-    }
-  `;
-}
-
-async function handleTodayPlanGenerate() {
-  const goalInput = document.getElementById("todayPlanGoalInput");
-  const goalText =
-    goalInput instanceof HTMLInputElement ? goalInput.value.trim() : "";
-  state.todayPlanState.goalText = goalText;
-  state.todayPlanState.generating = true;
-  state.todayPlanState.loading = true;
-  state.todayPlanState.error = "";
-  state.todayPlanState.unavailable = false;
-  state.todayPlanState.loadingMessage = "Generating plan preview...";
-  renderTodayPlanPanel();
-
-  const generationId = ++state.todayPlanGenerationSeq;
-  await loadTodayPlanDecisionAssist(true);
-  if (generationId !== state.todayPlanGenerationSeq) return;
-  state.todayPlanState.loadingMessage = "";
-  state.todayPlanState.lastApplyBatch = null;
-  renderTodayPlanPanel();
-}
-
-function handleTodayPlanToggleItem(todoId, checked) {
-  if (checked) {
-    state.todayPlanState.selectedTodoIds.add(todoId);
-  } else {
-    state.todayPlanState.selectedTodoIds.delete(todoId);
-  }
-  renderTodayPlanPanel();
-}
-
-async function handleTodayPlanDismissSuggestion(suggestionId) {
-  if (
-    state.todayPlanState.mode === "live" &&
-    state.todayPlanState.aiSuggestionId
-  ) {
-    try {
-      await apiCall(
-        `${API_URL}/ai/suggestions/${encodeURIComponent(state.todayPlanState.aiSuggestionId)}/dismiss`,
-        {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ suggestionId, dismissAll: true }),
-        },
-      );
-    } catch (error) {
-      console.error("Today plan dismiss failed:", error);
-    }
-    state.todayPlanState.aiSuggestionId = "";
-    state.todayPlanState.envelope = normalizeTodayPlanEnvelope({
-      surface: TODAY_PLAN_SURFACE,
-      must_abstain: false,
-      planPreview: { topN: 3, items: [] },
-      suggestions: [],
-    });
-    state.todayPlanState.selectedTodoIds = new Set();
-    state.todayPlanState.dismissedSuggestionIds = new Set();
-    renderTodayPlanPanel();
-    return;
-  }
-  state.todayPlanState.dismissedSuggestionIds.add(suggestionId);
-  renderTodayPlanPanel();
-}
-
-async function handleTodayPlanApplySelected() {
-  if (!state.todayPlanState.envelope) return;
-  const suggestionsToApply = getTodayPlanSelectedSuggestionCards();
-  if (!suggestionsToApply.length) return;
-
-  const affectedTodoIds = new Set(
-    suggestionsToApply
-      .map((suggestion) => String(suggestion.payload?.todoId || ""))
-      .filter(Boolean),
-  );
-  const todoSnapshots = {};
-  for (const todoId of affectedTodoIds) {
-    const todo = state.todos.find((entry) => String(entry.id) === todoId);
-    if (!todo) continue;
-    todoSnapshots[todoId] = deepClone(todo);
-  }
-  const notesDraftSnapshot = deepClone(state.todayPlanState.notesDraftByTodoId);
-  if (
-    state.todayPlanState.mode === "live" &&
-    state.todayPlanState.aiSuggestionId
-  ) {
-    try {
-      const response = await apiCall(
-        `${API_URL}/ai/suggestions/${encodeURIComponent(state.todayPlanState.aiSuggestionId)}/apply`,
-        {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            selectedTodoIds: Array.from(state.todayPlanState.selectedTodoIds),
-            confirmed: true,
-          }),
-        },
-      );
-      if (!response.ok) {
-        const data = await response.json().catch(() => ({}));
-        state.todayPlanState.error =
-          typeof data?.error === "string"
-            ? data.error
-            : "Could not apply selected suggestions.";
-        renderTodayPlanPanel();
-        return;
-      }
-      const data = await response.json().catch(() => ({}));
-      const updatedTodos = Array.isArray(data?.todos) ? data.todos : [];
-      for (const updatedTodo of updatedTodos) {
-        if (!updatedTodo?.id) continue;
-        const index = state.todos.findIndex(
-          (item) => String(item.id) === String(updatedTodo.id),
-        );
-        if (index >= 0) {
-          state.todos[index] = updatedTodo;
-        }
-      }
-      await loadTodayPlanDecisionAssist(false);
-    } catch (error) {
-      console.error("Today plan apply failed:", error);
-      state.todayPlanState.error = "Could not apply selected suggestions.";
-      renderTodayPlanPanel();
-      return;
-    }
-  } else {
-    for (const suggestion of suggestionsToApply) {
-      const todoId = String(suggestion.payload?.todoId || "");
-      const todo = state.todos.find((entry) => String(entry.id) === todoId);
-      if (!todo) continue;
-
-      if (suggestion.type === "set_priority") {
-        todo.priority = normalizePriorityValue(suggestion.payload?.priority);
-        continue;
-      }
-      if (suggestion.type === "set_due_date") {
-        const dueDate = String(suggestion.payload?.dueDateISO || "");
-        const parsed = new Date(dueDate);
-        if (!Number.isNaN(parsed.getTime())) {
-          todo.dueDate = parsed.toISOString();
-        }
-        continue;
-      }
-      if (suggestion.type === "split_subtasks") {
-        const subtasksRaw = Array.isArray(suggestion.payload?.subtasks)
-          ? suggestion.payload.subtasks
-          : [];
-        todo.subtasks = subtasksRaw.slice(0, 5).map((subtask, index) => ({
-          id: `local-today-plan-${suggestion.suggestionId}-${index + 1}`,
-          title: String(subtask?.title || "").slice(0, 200),
-          completed: false,
-          order: Number(subtask?.order) || index + 1,
-        }));
-        continue;
-      }
-      if (suggestion.type === "propose_next_action") {
-        state.todayPlanState.notesDraftByTodoId[todoId] = String(
-          suggestion.payload?.text || "",
-        ).slice(0, 500);
-      }
-    }
-  }
-
-  state.todayPlanState.lastApplyBatch = {
-    todoSnapshots,
-    notesDraftSnapshot,
-  };
-  renderTodos();
-}
-
-function handleTodayPlanUndoBatch() {
-  const batch = state.todayPlanState.lastApplyBatch;
-  if (!batch) return;
-  Object.entries(batch.todoSnapshots || {}).forEach(([todoId, snapshot]) => {
-    const index = state.todos.findIndex((entry) => String(entry.id) === todoId);
-    if (index === -1) return;
-    state.todos[index] = snapshot;
-  });
-  state.todayPlanState.notesDraftByTodoId = deepClone(
-    batch.notesDraftSnapshot || {},
-  );
-  state.todayPlanState.lastApplyBatch = null;
-  emitAiSuggestionUndoTelemetry({
-    surface: TODAY_PLAN_SURFACE,
-    aiSuggestionDbId: state.todayPlanState.aiSuggestionId,
-    suggestionId: state.todayPlanState.envelope?.requestId || "",
-    selectedTodoIdsCount: Object.keys(batch.todoSnapshots || {}).length,
-  });
-  renderTodos();
-}
-
-function bindTodayPlanHandlers() {
-  if (window.__todayPlanHandlersBound) {
-    return;
-  }
-  window.__todayPlanHandlersBound = true;
-
-  document.addEventListener("input", (event) => {
-    const target = event.target;
-    if (!(target instanceof HTMLElement)) return;
-    if (target.id !== "todayPlanGoalInput") return;
-    if (!(target instanceof HTMLInputElement)) return;
-    state.todayPlanState.goalText = target.value;
-  });
-
-  document.addEventListener("change", (event) => {
-    const target = event.target;
-    if (!(target instanceof HTMLElement)) return;
-    const action = target.getAttribute("data-today-plan-action");
-    if (action !== "toggle-item") return;
-    const todoId = String(target.getAttribute("data-today-plan-todo-id") || "");
-    if (!todoId || !(target instanceof HTMLInputElement)) return;
-    handleTodayPlanToggleItem(todoId, target.checked);
-  });
-
-  document.addEventListener("click", async (event) => {
-    const target = event.target;
-    if (!(target instanceof Element)) return;
-    const actionEl = target.closest("[data-today-plan-action]");
-    if (!(actionEl instanceof HTMLElement)) return;
-    const action = actionEl.getAttribute("data-today-plan-action");
-    if (action === "toggle-item" && actionEl instanceof HTMLInputElement) {
-      const todoId = String(
-        actionEl.getAttribute("data-today-plan-todo-id") || "",
-      );
-      if (!todoId) return;
-      handleTodayPlanToggleItem(todoId, actionEl.checked);
-      return;
-    }
-    if (action === "generate") {
-      await handleTodayPlanGenerate();
-      return;
-    }
-    if (action === "dismiss-suggestion") {
-      const suggestionId = String(
-        actionEl.getAttribute("data-today-plan-suggestion-id") || "",
-      );
-      if (!suggestionId) return;
-      await handleTodayPlanDismissSuggestion(suggestionId);
-      return;
-    }
-    if (action === "apply-selected") {
-      await handleTodayPlanApplySelected();
-      return;
-    }
-    if (action === "undo") {
-      handleTodayPlanUndoBatch();
-    }
-  });
 }
 
 // ========== PHASE B: PRIORITY, NOTES, SUBTASKS ==========
@@ -3988,333 +1036,6 @@ async function aiBreakdownTodo(todoId, force = false) {
   } catch (error) {
     console.error("AI breakdown error:", error);
     showMessage("todosMessage", "Failed to generate subtasks", "error");
-  }
-}
-
-// ========== PHASE A: DRAG & DROP FUNCTIONALITY ==========
-let draggedOverTodoId = null;
-let draggedHeadingId = null;
-
-function resolveDragRow(event, selector, fallbackElement = null) {
-  const candidateTargets = [];
-  if (fallbackElement instanceof Element) {
-    candidateTargets.push(fallbackElement);
-  }
-  const target = event?.target;
-  if (target instanceof Element) {
-    candidateTargets.push(target);
-  }
-  const currentTarget = event?.currentTarget;
-  if (currentTarget instanceof Element) {
-    candidateTargets.push(currentTarget);
-  }
-
-  for (const candidate of candidateTargets) {
-    if (candidate.matches(selector)) {
-      return candidate instanceof HTMLElement ? candidate : null;
-    }
-    const closest = candidate.closest(selector);
-    if (closest instanceof HTMLElement) {
-      return closest;
-    }
-  }
-
-  const clientX = Number(event?.clientX);
-  const clientY = Number(event?.clientY);
-  if (Number.isFinite(clientX) && Number.isFinite(clientY)) {
-    const pointTarget = document.elementFromPoint(clientX, clientY);
-    if (pointTarget instanceof Element) {
-      if (pointTarget.matches(selector)) {
-        return pointTarget instanceof HTMLElement ? pointTarget : null;
-      }
-      const closest = pointTarget.closest(selector);
-      if (closest instanceof HTMLElement) {
-        return closest;
-      }
-    }
-  }
-
-  return null;
-}
-
-function getTodoRowFromDragEvent(event, fallbackElement = null) {
-  return resolveDragRow(event, ".todo-item[data-todo-id]", fallbackElement);
-}
-
-function getHeadingRowFromDragEvent(event, fallbackElement = null) {
-  return resolveDragRow(
-    event,
-    ".todo-heading-divider[data-heading-id]",
-    fallbackElement,
-  );
-}
-
-function handleDragStart(e, rowElement = null) {
-  const row = getTodoRowFromDragEvent(e, rowElement);
-  if (!row) return;
-  state.draggedTodoId = row.dataset.todoId;
-  draggedHeadingId = null;
-  row.classList.add("dragging");
-  if (e.dataTransfer) {
-    e.dataTransfer.effectAllowed = "move";
-    e.dataTransfer.setData("text/plain", state.draggedTodoId || "");
-  }
-}
-
-function handleDragOver(e, rowElement = null) {
-  if (draggedHeadingId) {
-    const row = getTodoRowFromDragEvent(e, rowElement);
-    if (!row) return;
-    const todoId = String(row.dataset.todoId || "");
-    if (!todoId) return;
-    e.preventDefault();
-    clearHeadingDragState();
-    const bounds = row.getBoundingClientRect();
-    const dropPosition =
-      e.clientY > bounds.top + bounds.height / 2 ? "after" : "before";
-    row.dataset.headingDropPosition = dropPosition;
-    row.classList.add("todo-item--heading-drop-target");
-    row.classList.add(
-      dropPosition === "after"
-        ? "todo-item--heading-drop-after"
-        : "todo-item--heading-drop-before",
-    );
-    if (e.dataTransfer) {
-      e.dataTransfer.dropEffect = "move";
-    }
-    return;
-  }
-
-  e.preventDefault();
-  if (e.dataTransfer) {
-    e.dataTransfer.dropEffect = "move";
-  }
-
-  const row = getTodoRowFromDragEvent(e, rowElement);
-  if (!row) return;
-  document.querySelectorAll(".todo-item").forEach((item) => {
-    if (item === row) return;
-    delete item.dataset.todoDropPosition;
-    item.classList.remove("drag-over");
-  });
-  const todoId = row?.dataset.todoId || "";
-  if (todoId !== state.draggedTodoId) {
-    const bounds = row.getBoundingClientRect();
-    const dropPosition =
-      e.clientY > bounds.top + bounds.height / 2 ? "after" : "before";
-    row.dataset.todoDropPosition = dropPosition;
-    row?.classList.add("drag-over");
-    draggedOverTodoId = todoId;
-  }
-}
-
-function handleDrop(e, rowElement = null) {
-  if (draggedHeadingId) {
-    e.preventDefault();
-    e.stopPropagation();
-    const row = getTodoRowFromDragEvent(e, rowElement);
-    if (!row) return;
-    const targetTodoId = String(row.dataset.todoId || "");
-    const dropPosition =
-      row.dataset.headingDropPosition === "after" ? "after" : "before";
-    const headingDropTarget = getHeadingDropTargetFromTodo(
-      targetTodoId,
-      dropPosition,
-    );
-    if (headingDropTarget) {
-      reorderProjectHeadings(
-        draggedHeadingId,
-        headingDropTarget.targetId,
-        headingDropTarget.placement,
-      );
-    }
-    clearHeadingDragState();
-    draggedHeadingId = null;
-    return;
-  }
-
-  e.preventDefault();
-  e.stopPropagation();
-
-  const row = getTodoRowFromDragEvent(e, rowElement);
-  const dropTargetId = row?.dataset.todoId || "";
-  const placement =
-    row?.dataset.todoDropPosition === "after" ? "after" : "before";
-  if (row) {
-    delete row.dataset.todoDropPosition;
-  }
-  row?.classList.remove("drag-over");
-
-  if (
-    state.draggedTodoId &&
-    dropTargetId &&
-    state.draggedTodoId !== dropTargetId
-  ) {
-    const selectedProject = getSelectedProjectKey();
-    const targetTodo =
-      state.todos.find((todo) => todo.id === dropTargetId) || null;
-    const nextHeadingId = selectedProject
-      ? String(targetTodo?.headingId || "")
-      : null;
-    reorderTodos(state.draggedTodoId, dropTargetId, {
-      nextHeadingId: selectedProject ? nextHeadingId || null : undefined,
-      placement,
-    });
-  }
-}
-
-function handleDragEnd(e, rowElement = null) {
-  const row = getTodoRowFromDragEvent(e, rowElement);
-  row?.classList.remove("dragging");
-  document.querySelectorAll(".todo-item").forEach((item) => {
-    item.classList.remove("drag-over");
-    delete item.dataset.todoDropPosition;
-  });
-  clearHeadingDragState();
-  state.draggedTodoId = null;
-  draggedOverTodoId = null;
-}
-
-function clearHeadingDragState() {
-  document.querySelectorAll(".todo-heading-divider").forEach((row) => {
-    row.classList.remove(
-      "todo-heading-divider--dragging",
-      "todo-heading-divider--drag-over-before",
-      "todo-heading-divider--drag-over-after",
-    );
-    delete row.dataset.headingDropPosition;
-  });
-  document.querySelectorAll(".todo-item").forEach((row) => {
-    row.classList.remove(
-      "todo-item--heading-drop-target",
-      "todo-item--heading-drop-before",
-      "todo-item--heading-drop-after",
-    );
-    delete row.dataset.headingDropPosition;
-  });
-}
-
-function handleHeadingDragStart(e, rowElement = null) {
-  const row = getHeadingRowFromDragEvent(e, rowElement);
-  if (!row) return;
-  const headingId = row.dataset.headingId || "";
-  if (!headingId) return;
-
-  draggedHeadingId = headingId;
-  state.draggedTodoId = null;
-  row.classList.add("todo-heading-divider--dragging");
-  if (e.dataTransfer) {
-    e.dataTransfer.effectAllowed = "move";
-    e.dataTransfer.setData("text/plain", headingId);
-  }
-}
-
-function handleHeadingDragOver(e, rowElement = null) {
-  const row = getHeadingRowFromDragEvent(e, rowElement);
-  if (!row) return;
-  const targetHeadingId = row.dataset.headingId || "";
-
-  if (draggedHeadingId) {
-    if (!targetHeadingId || targetHeadingId === draggedHeadingId) return;
-    e.preventDefault();
-    clearHeadingDragState();
-    const bounds = row.getBoundingClientRect();
-    const dropPosition =
-      e.clientY > bounds.top + bounds.height / 2 ? "after" : "before";
-    row.dataset.headingDropPosition = dropPosition;
-    row.classList.add(
-      dropPosition === "after"
-        ? "todo-heading-divider--drag-over-after"
-        : "todo-heading-divider--drag-over-before",
-    );
-    if (e.dataTransfer) {
-      e.dataTransfer.dropEffect = "move";
-    }
-    return;
-  }
-
-  if (!state.draggedTodoId || !targetHeadingId) return;
-  e.preventDefault();
-  clearHeadingDragState();
-  row.classList.add("todo-heading-divider--drag-over-before");
-  if (e.dataTransfer) {
-    e.dataTransfer.dropEffect = "move";
-  }
-}
-
-function getFirstTodoIdInHeading(headingId, excludeTodoId = null) {
-  const selectedProject = getSelectedProjectKey();
-  const normalizedProject = normalizeProjectPath(selectedProject);
-  const candidates = [...state.todos]
-    .filter((todo) => {
-      const todoProject = normalizeProjectPath(todo.category || "");
-      if (normalizedProject && todoProject !== normalizedProject) return false;
-      if (excludeTodoId && String(todo.id) === String(excludeTodoId))
-        return false;
-      return String(todo.headingId || "") === String(headingId || "");
-    })
-    .sort((a, b) => (a.order || 0) - (b.order || 0));
-  return candidates[0]?.id || null;
-}
-
-function handleHeadingDrop(e, rowElement = null) {
-  e.preventDefault();
-  e.stopPropagation();
-  const row = getHeadingRowFromDragEvent(e, rowElement);
-  if (!row) return;
-  const targetHeadingId = row.dataset.headingId || "";
-  if (!targetHeadingId) return;
-
-  if (draggedHeadingId) {
-    const dropPosition =
-      row.dataset.headingDropPosition === "after" ? "after" : "before";
-    if (draggedHeadingId !== targetHeadingId) {
-      reorderProjectHeadings(draggedHeadingId, targetHeadingId, dropPosition);
-    }
-    clearHeadingDragState();
-    draggedHeadingId = null;
-    return;
-  }
-
-  if (!state.draggedTodoId) return;
-  const firstTodoId = getFirstTodoIdInHeading(
-    targetHeadingId,
-    state.draggedTodoId,
-  );
-  if (firstTodoId) {
-    reorderTodos(state.draggedTodoId, firstTodoId, {
-      nextHeadingId: targetHeadingId,
-    });
-  } else {
-    moveTodoToHeading(state.draggedTodoId, targetHeadingId);
-  }
-  clearHeadingDragState();
-}
-
-function handleHeadingDragEnd() {
-  clearHeadingDragState();
-  draggedHeadingId = null;
-}
-
-// ========== PHASE A: KEYBOARD SHORTCUTS ==========
-function toggleShortcuts() {
-  const overlay = document.getElementById("shortcutsOverlay");
-  if (!overlay) return;
-  const isNowOpen = !overlay.classList.contains("active");
-  overlay.classList.toggle("active");
-  if (isNowOpen) {
-    DialogManager.open("shortcuts", overlay, {
-      onEscape: toggleShortcuts,
-      backdrop: false,
-    });
-  } else {
-    DialogManager.close("shortcuts");
-  }
-}
-
-function closeShortcutsOverlay(event) {
-  if (event.target.id === "shortcutsOverlay") {
-    toggleShortcuts();
   }
 }
 
@@ -4527,143 +1248,6 @@ document.addEventListener("keydown", function (e) {
   }
 });
 
-// Load admin users
-async function loadAdminUsers() {
-  hideMessage("adminMessage");
-
-  try {
-    const response = await apiCall(`${API_URL}/admin/users`);
-    if (response && response.ok) {
-      state.users = await response.json();
-      renderAdminUsers();
-    } else {
-      const data = await response.json();
-      showMessage(
-        "adminMessage",
-        data.error || "Failed to load users",
-        "error",
-      );
-    }
-  } catch (error) {
-    showMessage("adminMessage", "Network error. Please try again.", "error");
-    console.error("Load users error:", error);
-  }
-}
-
-// Render admin users
-function renderAdminUsers() {
-  const container = document.getElementById("adminContent");
-
-  container.innerHTML = `
-                <table class="users-table">
-                    <thead>
-                        <tr>
-                            <th>Email</th>
-                            <th>Name</th>
-                            <th>Role</th>
-                            <th>Verified</th>
-                            <th>Joined</th>
-                            <th>Actions</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        ${state.users
-                          .map(
-                            (user) => `
-                            <tr>
-                                <td>${escapeHtml(user.email)}</td>
-                                <td>${escapeHtml(user.name || "-")}</td>
-                                <td><span class="role-badge ${user.role}">${user.role}</span></td>
-                                <td>${user.isVerified ? "✓" : "✗"}</td>
-                                <td>${new Date(user.createdAt).toLocaleDateString()}</td>
-                                <td>
-                                    ${
-                                      user.id !== state.currentUser.id
-                                        ? `
-                                        ${
-                                          user.role === "user"
-                                            ? `
-                                            <button class="action-btn promote" data-onclick="changeUserRole('${user.id}', 'admin')">Make Admin</button>
-                                        `
-                                            : `
-                                            <button class="action-btn demote" data-onclick="changeUserRole('${user.id}', 'user')">Remove Admin</button>
-                                        `
-                                        }
-                                        <button class="action-btn delete" data-onclick="deleteUser('${user.id}')">Delete</button>
-                                    `
-                                        : "<em>You</em>"
-                                    }
-                                </td>
-                            </tr>
-                        `,
-                          )
-                          .join("")}
-                    </tbody>
-                </table>
-            `;
-}
-
-// Change user role
-async function changeUserRole(userId, role) {
-  hideMessage("adminMessage");
-
-  try {
-    const response = await apiCall(`${API_URL}/admin/users/${userId}/role`, {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ role }),
-    });
-
-    if (response && response.ok) {
-      showMessage("adminMessage", `User role updated to ${role}`, "success");
-      loadAdminUsers();
-    } else {
-      const data = response ? await parseApiBody(response) : {};
-      showMessage(
-        "adminMessage",
-        data.error || "Failed to update role",
-        "error",
-      );
-    }
-  } catch (error) {
-    showMessage("adminMessage", "Network error. Please try again.", "error");
-    console.error("Change role error:", error);
-  }
-}
-
-// Delete user
-async function deleteUser(userId) {
-  if (
-    !(await showConfirmDialog(
-      "Are you sure you want to delete this user? This action cannot be undone.",
-    ))
-  )
-    return;
-
-  hideMessage("adminMessage");
-
-  try {
-    const response = await apiCall(`${API_URL}/admin/users/${userId}`, {
-      method: "DELETE",
-    });
-
-    if (response && response.ok) {
-      showMessage("adminMessage", "User deleted successfully", "success");
-      loadAdminUsers();
-    } else {
-      const data = response ? await parseApiBody(response) : {};
-      showMessage(
-        "adminMessage",
-        data.error || "Failed to delete user",
-        "error",
-      );
-    }
-  } catch (error) {
-    showMessage("adminMessage", "Network error. Please try again.", "error");
-    console.error("Delete user error:", error);
-  }
-}
-
 // Switch view
 function switchView(view, triggerEl = null) {
   const requestedView = view === "profile" ? "settings" : view;
@@ -4742,52 +1326,6 @@ function shouldIgnoreTodoDrawerOpen(target) {
   return !!target.closest(
     "input, button, select, textarea, a, label, [data-onclick], [data-onchange], .drag-handle, .todo-inline-actions, .subtasks-section, .todo-kebab, .todo-kebab-menu",
   );
-}
-
-function bindCommandPaletteHandlers() {
-  if (window.__commandPaletteHandlersBound) {
-    return;
-  }
-  window.__commandPaletteHandlersBound = true;
-
-  document.addEventListener("click", (event) => {
-    const target = event.target;
-    if (!(target instanceof Element)) return;
-
-    const overlay = target.closest("#commandPaletteOverlay");
-    if (
-      overlay instanceof HTMLElement &&
-      target.id === "commandPaletteOverlay"
-    ) {
-      closeCommandPalette({ restoreFocus: true });
-      return;
-    }
-
-    if (!state.isCommandPaletteOpen) return;
-
-    const option = target.closest("[data-command-id]");
-    if (!(option instanceof HTMLElement)) return;
-    event.preventDefault();
-    const itemIndex = Number.parseInt(
-      option.getAttribute("data-command-index") || "-1",
-      10,
-    );
-    if (itemIndex < 0) return;
-    executeCommandPaletteItem(
-      state.commandPaletteSelectableItems[itemIndex],
-      option,
-    );
-  });
-
-  document.addEventListener("input", (event) => {
-    const target = event.target;
-    if (!(target instanceof HTMLElement)) return;
-    if (target.id !== "commandPaletteInput") return;
-    state.commandPaletteQuery =
-      target instanceof HTMLInputElement ? target.value : String(target.value);
-    state.commandPaletteIndex = 0;
-    renderCommandPalette();
-  });
 }
 
 function bindCriticalHandlers() {
@@ -5168,8 +1706,8 @@ function bindDeclarativeHandlers() {
   // authUi cross-module hooks
   hooks.switchView = switchView;
   hooks.closeCommandPalette = closeCommandPalette;
-  hooks.resetOnCreateAssistState = resetOnCreateAssistState;
-  hooks.resetTodayPlanState = resetTodayPlanState;
+  hooks.resetOnCreateAssistState = OnCreateAssist.resetOnCreateAssistState;
+  hooks.resetTodayPlanState = TodayPlan.resetTodayPlanState;
   hooks.clearPlanDraftState = clearPlanDraftState;
   hooks.setTodosViewBodyState = setTodosViewBodyState;
   hooks.setSettingsPaneVisible = setSettingsPaneVisible;
@@ -5194,7 +1732,7 @@ function bindDeclarativeHandlers() {
   hooks.ensureTodosShellActive = ensureTodosShellActive;
   // filterLogic → render sub-hooks
   hooks.renderProjectsRail = renderProjectsRail;
-  hooks.renderTodayPlanPanel = renderTodayPlanPanel;
+  hooks.renderTodayPlanPanel = TodayPlan.renderTodayPlanPanel;
   hooks.clearHomeFocusDashboard = clearHomeFocusDashboard;
   hooks.renderHomeDashboard = renderHomeDashboard;
   hooks.updateBulkActionsVisibility = updateBulkActionsVisibility;
@@ -5210,24 +1748,38 @@ function bindDeclarativeHandlers() {
   // Additional hooks needed by domain modules
   hooks.buildUrl = ApiClientModule.buildUrl;
   hooks.buildHomeTileListByKey = buildHomeTileListByKey;
+  // New module hooks (task 149)
+  hooks.hideMessage = hideMessage;
+  hooks.moveTodoToHeading = moveTodoToHeading;
+  hooks.reorderProjectHeadings = reorderProjectHeadings;
+  hooks.openTodoDrawer = openTodoDrawer;
+  hooks.clearHomeListDrilldown = clearHomeListDrilldown;
+  hooks.impactRankForSurface = impactRankForSurface;
+  hooks.ON_CREATE_SURFACE = ON_CREATE_SURFACE;
+  hooks.TODAY_PLAN_SURFACE = TODAY_PLAN_SURFACE;
+  hooks.isKnownSuggestionType = isKnownSuggestionType;
+  hooks.initializeDrawerDraft = initializeDrawerDraft;
+  hooks.setDrawerSaveState = setDrawerSaveState;
+  hooks.getTodoDrawerElements = getTodoDrawerElements;
+  hooks.escapeSelectorValue = TaskDrawerAssist.escapeSelectorValue;
   hooks.buildIcsContentForTodos = buildIcsContentForTodos;
   hooks.buildIcsFilename = buildIcsFilename;
-  hooks.clearOnCreateDismissed = clearOnCreateDismissed;
+  hooks.clearOnCreateDismissed = OnCreateAssist.clearOnCreateDismissed;
   hooks.closeTaskComposer = closeTaskComposer;
   hooks.filterTodosList = filterTodosList;
   hooks.getHomeDrilldownLabel = getHomeDrilldownLabel;
   hooks.getProjectHeadings = getProjectHeadings;
-  hooks.getTodoById = getTodoById;
+  hooks.getTodoById = TaskDrawerAssist.getTodoById;
   hooks.getVisibleTodos = getVisibleTodos;
   hooks.isInternalCategoryPath = isInternalCategoryPath;
-  hooks.loadOnCreateDecisionAssist = loadOnCreateDecisionAssist;
+  hooks.loadOnCreateDecisionAssist = OnCreateAssist.loadOnCreateDecisionAssist;
   hooks.openTaskComposer = openTaskComposer;
   hooks.processQuickEntryNaturalDate = processQuickEntryNaturalDate;
   hooks.renderHomeFocusDashboard = renderHomeFocusDashboard;
-  hooks.renderOnCreateAssistRow = renderOnCreateAssistRow;
+  hooks.renderOnCreateAssistRow = OnCreateAssist.renderOnCreateAssistRow;
   hooks.renderProjectOptions = renderProjectOptions;
   hooks.renderSubtasks = renderSubtasks;
-  hooks.renderTodoChips = renderTodoChips;
+  hooks.renderTodoChips = TaskDrawerAssist.renderTodoChips;
   hooks.resetQuickEntryNaturalDueState = resetQuickEntryNaturalDueState;
   hooks.selectProjectFromRail = selectProjectFromRail;
   hooks.selectWorkspaceView = selectWorkspaceView;
@@ -5322,14 +1874,14 @@ window.moveTodoToProject = moveTodoToProject;
 window.moveTodoToHeading = moveTodoToHeading;
 window.moveProjectHeading = moveProjectHeading;
 // Drag and drop
-window.handleDragStart = handleDragStart;
-window.handleDragOver = handleDragOver;
-window.handleDrop = handleDrop;
-window.handleDragEnd = handleDragEnd;
-window.handleHeadingDragStart = handleHeadingDragStart;
-window.handleHeadingDragOver = handleHeadingDragOver;
-window.handleHeadingDrop = handleHeadingDrop;
-window.handleHeadingDragEnd = handleHeadingDragEnd;
+window.handleDragStart = DragDrop.handleDragStart;
+window.handleDragOver = DragDrop.handleDragOver;
+window.handleDrop = DragDrop.handleDrop;
+window.handleDragEnd = DragDrop.handleDragEnd;
+window.handleHeadingDragStart = DragDrop.handleHeadingDragStart;
+window.handleHeadingDragOver = DragDrop.handleHeadingDragOver;
+window.handleHeadingDrop = DragDrop.handleHeadingDrop;
+window.handleHeadingDragEnd = DragDrop.handleHeadingDragEnd;
 // AI critique / plan
 window.applyCritiqueSuggestion = applyCritiqueSuggestion;
 window.applyCritiqueSuggestionMode = applyCritiqueSuggestionMode;
@@ -5368,8 +1920,8 @@ function init() {
   bindCommandPaletteHandlers();
   bindTaskComposerHandlers();
   bindDockHandlers();
-  bindOnCreateAssistHandlers();
-  bindTodayPlanHandlers();
+  OnCreateAssist.bindOnCreateAssistHandlers();
+  TodayPlan.bindTodayPlanHandlers();
   bindQuickEntryNaturalDateHandlers();
 
   // Check for reset token in URL
@@ -5404,7 +1956,7 @@ function init() {
   } else {
     setAuthState(AUTH_STATE.UNAUTHENTICATED);
   }
-  renderOnCreateAssistRow();
+  OnCreateAssist.renderOnCreateAssistRow();
   setQuickEntryPropertiesOpen(readStoredQuickEntryPropertiesOpenState(), {
     persist: false,
   });
